+++ smb/dcerpc/dcerpc.go
@@ -47,17 +47,18 @@ import (
 var log = golog.Get("github.com/jfjallid/go-smb/smb/dcerpc")
 
 var (
-	MSRPCSvcCtlPipe                       = "svcctl"
-	MSRPCUuidSvcCtl                       = "367ABB81-9844-35F1-AD32-98F038001003"
-	MSRPCSvcCtlMajorVersion uint16        = 2
-	MSRPCSvcCtlMinorVersion uint16        = 0
-	MSRPCUuidSrvSvc                       = "4B324FC8-1670-01D3-1278-5A47BF6EE188"
-	MSRPCSrvSvcMajorVersion uint16        = 3
-	MSRPCSrvSvcMinorVersion uint16        = 0
-	MSRPCUuidNdr                          = "8a885d04-1ceb-11c9-9fe8-08002b104860" // NDR Transfer Syntax version 2.0
-	re                      regexp.Regexp = *regexp.MustCompile(`([\dA-Fa-f]{8})-([\dA-Fa-f]{4})-([\dA-Fa-f]{4})-([\dA-Fa-f]{4})-([\dA-Fa-f]{4})([\dA-Fa-f]{8})`)
-	ContextItemLen                        = 44
-	ContextResItemLen                     = 24
+	MSRPCSvcCtlPipe                          = "svcctl"
+	MSRPCUuidSvcCtl                          = "367ABB81-9844-35F1-AD32-98F038001003"
+	MSRPCSvcCtlMajorVersion uint16           = 2
+	MSRPCSvcCtlMinorVersion uint16           = 0
+	MSRPCUuidSrvSvc                          = "4B324FC8-1670-01D3-1278-5A47BF6EE188"
+	MSRPCSrvSvcMajorVersion uint16           = 3
+	MSRPCSrvSvcMinorVersion uint16           = 0
+	MSRPCUuidNdr                             = "8a885d04-1ceb-11c9-9fe8-08002b104860" // NDR Transfer Syntax version 2.0
+	re                      regexp.Regexp    = *regexp.MustCompile(`([\dA-Fa-f]{8})-([\dA-Fa-f]{4})-([\dA-Fa-f]{4})-([\dA-Fa-f]{4})-([\dA-Fa-f]{4})([\dA-Fa-f]{8})`)
+	ContextItemLen                           = 44
+	ContextResItemLen                        = 24
+	le                      binary.ByteOrder = binary.LittleEndian
 )
 
 // MSRPC Packet Types
@@ -67,261 +68,6 @@ const (
 	PacketTypeBind     uint8 = 11
 	PacketTypeBindAck  uint8 = 12
 )
-
-// MSRPC Server Service (srvsvc) Operations
-const (
-	SrvSvcOpNetrSessionEnum  uint16 = 12
-	SrvSvcOpNetShareEnumAll  uint16 = 15
-	SrvSvcOpNetServerGetInfo uint16 = 21
-)
-
-// MS-SCMR Operations OP Codes
-const (
-	SvcCtlRCloseServiceHandle   uint16 = 0
-	SvcCtlRControlService       uint16 = 1
-	SvcCtlRDeleteService        uint16 = 2
-	SvcCtlRQueryServiceStatus   uint16 = 6
-	SvcCtlRChangeServiceConfigW uint16 = 11
-	SvcCtlRCreateServiceW       uint16 = 12
-	SvcCtlROpenSCManagerW       uint16 = 15
-	SvcCtlROpenServiceW         uint16 = 16
-	SvcCtlRQueryServiceConfigW  uint16 = 17
-	SvcCtlRStartServiceW        uint16 = 19
-)
-
-// MS-SCMR (svcctl) Section 2.2.15 ServiceType merged with Section 2.2.47 dwServiceType
-const (
-	ServiceKernelDriver       uint32 = 0x00000001
-	ServiceFileSystemDriver   uint32 = 0x00000002
-	ServiceWin32OwnProcess    uint32 = 0x00000010
-	ServiceWin32ShareProcess  uint32 = 0x00000020
-	ServiceInteractiveProcess uint32 = 0x00000100
-)
-
-var ServiceTypeStatusMap = map[uint32]string{
-	ServiceKernelDriver:       "SERVICE_KERNEL_DRIVER",
-	ServiceFileSystemDriver:   "SERVICE_FILE_SYSTEM_DRIVER",
-	ServiceWin32OwnProcess:    "SERIVCE_WIN32_OWN_PROCESS",
-	ServiceWin32ShareProcess:  "SERVICE_WIN32_SHARE_PROCESS",
-	ServiceInteractiveProcess: "SERVICE_INTERACTIVE_PROCESS",
-}
-
-// MS-SCMR (svcctl) Table 2.2.15 StartType
-const (
-	ServiceBootStart   uint32 = 0x00000000
-	ServiceSystemStart uint32 = 0x00000001
-	ServiceAutoStart   uint32 = 0x00000002
-	ServiceDemandStart uint32 = 0x00000003
-	ServiceDisabled    uint32 = 0x00000004
-)
-
-var StartTypeStatusMap = map[uint32]string{
-	ServiceBootStart:   "SERVICE_BOOT_START",
-	ServiceSystemStart: "SERVICE_SYSTEM_START",
-	ServiceAutoStart:   "SERIVCE_AUTO_START",
-	ServiceDemandStart: "SERVICE_DEMAND_START",
-	ServiceDisabled:    "SERVICE_DISABLED",
-}
-
-// MS-SCMR (svcctl) Table 2.2.15 ErrorControl
-const (
-	ServiceErrorIgnore   uint32 = 0x00000000
-	ServiceErrorNormal   uint32 = 0x00000001
-	ServiceErrorSevere   uint32 = 0x00000002
-	ServiceErrorCritical uint32 = 0x00000003
-)
-
-var ErrorControlStatusMap = map[uint32]string{
-	ServiceErrorIgnore:   "SERVICE_ERROR_IGNORE",
-	ServiceErrorNormal:   "SERVICE_ERROR_NORMAL",
-	ServiceErrorSevere:   "SERIVCE_ERROR_SEVERE",
-	ServiceErrorCritical: "SERVICE_ERROR_CRITICAL",
-}
-
-// MS-SCMR (svcctl) Table 3.1.4
-const (
-	ServiceAllAccess           uint32 = 0x000F01FF //In addition to all access rights in this table, SERVICE_ALL_ACCESS includes Delete (DE), Read Control (RC), Write DACL (WD), and Write Owner (WO) access, as specified in ACCESS_MASK (section 2.4.3) of [MS-DTYP].
-	ServiceChangeConfig        uint32 = 0x00000002 //Required to change the configuration of a service.
-	ServiceEnumerateDependents uint32 = 0x00000008 //Required to enumerate the services installed on the server.
-	ServiceInterrogate         uint32 = 0x00000080 //Required to request immediate status from the service.
-	ServicePauseContinue       uint32 = 0x00000040 //Required to pause or continue the service.
-	ServiceQueryConfig         uint32 = 0x00000001 //Required to query the service configuration.
-	ServiceQueryStatus         uint32 = 0x00000004 //Required to request the service status.
-	ServiceStart               uint32 = 0x00000010 //Required to start the service.
-	ServiceStop                uint32 = 0x00000020 //Required to stop the service
-	ServiceUserDefinedControl  uint32 = 0x00000100 //Required to specify a user-defined control code.
-	ServiceSetStatus           uint32 = 0x00008000 //Required for a service to set its status.
-
-	SCManagerLock             uint32 = 0x00000008 //Required to lock the SCM database.
-	SCManagerCreateService    uint32 = 0x00000002 //Required for a service to be created.
-	SCManagerEnumerateService uint32 = 0x00000004 //Required to enumerate a service.
-	SCManagerConnect          uint32 = 0x00000001 //Required to connect to the SCM.
-	SCManagerQueryLockStatus  uint32 = 0x00000010 //Required to query the lock status of the SCM database.
-	SCManagerModifyBootConfig uint32 = 0x0020     //Required to call the RNotifyBootConfigStatus method.
-)
-
-// MS-SCMR Section 2.2.47 dwCurrentState
-const (
-	ServiceContinuePending uint32 = 0x00000005
-	ServicePausePending    uint32 = 0x00000006
-	ServicePaused          uint32 = 0x00000007
-	ServiceRunning         uint32 = 0x00000004
-	ServiceStartPending    uint32 = 0x00000002
-	ServiceStopPending     uint32 = 0x00000003
-	ServiceStopped         uint32 = 0x00000001
-)
-
-var ServiceStatusMap = map[uint32]string{
-	ServiceContinuePending: "SERVICE_CONTINUE_PENDING",
-	ServicePausePending:    "SERVICE_PAUSE_PENDING",
-	ServicePaused:          "SERVICE_PAUSED",
-	ServiceRunning:         "SERVICE_RUNNING",
-	ServiceStartPending:    "SERVICE_START_PENDING",
-	ServiceStopPending:     "SERVICE_STOP_PENDING",
-	ServiceStopped:         "SERVICE_STOPPED",
-}
-
-// MS-SCMR Section 3.1.4.2 dwControl
-const (
-	ServiceControlContinue       uint32 = 0x00000003
-	ServiceControlInterrogate    uint32 = 0x00000004
-	ServiceControlNetbindadd     uint32 = 0x00000007
-	ServiceControlNetbinddisable uint32 = 0x0000000A
-	ServiceControlNetbindenable  uint32 = 0x00000009
-	ServiceControlNetbindremove  uint32 = 0x00000008
-	ServiceControlParamChange    uint32 = 0x00000006
-	ServiceControlPause          uint32 = 0x00000002
-	ServiceControlStop           uint32 = 0x00000001
-)
-
-// MS-SCMR Section 3.1.4.11 RChangeServiceConfigW
-const (
-	ServiceNoChange uint32 = 0xffffffff
-)
-
-// MS-SCMR Response codes from multiple sections: 3.1.4.2, 3.1.4.11, 3.1.4.17, 3.1.4.19
-const (
-	ErrorSuccess                    uint32 = 0  // Successfully started the service
-	ErrorFileNotFound               uint32 = 2  // The system cannot find the file specified.
-	ErrorPathNotFound               uint32 = 3  // The system cannot find the path specified.
-	ErrorAccessDenied               uint32 = 5  // The SERVICE_START access right had not been granted to the caller when the RPC context handle to the service record was created.
-	ErrorInvalidHandle              uint32 = 6  // The handle is no longer valid.
-	ErrorInvalidParameter           uint32 = 87 // A parameter that was specified is invalid.
-	ErrorInsufficientBuffer         uint32 = 122
-	ErrorDependentServicesRunning   uint32 = 1051
-	ErrorInvalidServiceControl      uint32 = 1052
-	ErrorServiceRequestTimeout      uint32 = 1053 // The process for the service was started, but it did not respond within an implementation-specific time-out.
-	ErrorServiceNoThread            uint32 = 1054 // A thread could not be created for the service.
-	ErrorServiceDatabaseLocked      uint32 = 1055 // The service database is locked by the call to the BlockServiceDatabase method.
-	ErrorServiceAlreadyRunning      uint32 = 1056 // The ServiceStatus.dwCurrentState in the service record is not set to SERVICE_STOPPED.
-	ErrorInvalidServiceAccount      uint32 = 1057 // The user account name specified in the lpServiceStartName parameter does not exist.
-	ErrorServiceDisabled            uint32 = 1058 // The service cannot be started because the Start field in the service record is set to SERVICE_DISABLED.
-	ErrorCircularDependency         uint32 = 1059 // A circular dependency was specified.
-	ErrorServiceDoesNotExist        uint32 = 1060 // The service record with a specified display name does not exist in the SCM database
-	ErrorServiceCannotAcceptControl uint32 = 1061
-	ErrorServiceNotActive           uint32 = 1062
-	ErrorServiceDependencyFail      uint32 = 1068 // The specified service depends on another service that has failed to start.
-	ErrorServiceLogonFailed         uint32 = 1069 // The service did not start due to a logon failure.
-	ErrorServiceMarkedForDelete     uint32 = 1072 // The RDeleteService method has been called for the service record identified by the hService parameter.
-	ErrorServiceDependencyDeleted   uint32 = 1075 // The specified service depends on a service that does not exist or has been marked for deletion.
-	ErrorDuplicateServiceName       uint32 = 1078 // The lpDisplayName matches either the ServiceName or the DisplayName of another service record in the service control manager database.
-	ErrorShutdownInProgress         uint32 = 1115 // The system is shutting down.
-)
-
-var ServiceResponseCodeMap = map[uint32]error{
-	ErrorSuccess:                    fmt.Errorf("Successfully started the service"),
-	ErrorFileNotFound:               fmt.Errorf("The system cannot find the file specified."),
-	ErrorPathNotFound:               fmt.Errorf("The system cannot find the path specified."),
-	ErrorAccessDenied:               fmt.Errorf("ERROR_ACCESS_DENIED"),
-	ErrorInvalidHandle:              fmt.Errorf("The handle is no longer valid."),
-	ErrorInvalidParameter:           fmt.Errorf("A parameter that was specified is invalid."),
-	ErrorInsufficientBuffer:         fmt.Errorf("ERROR_INSUFFICIENT_BUFFER"),
-	ErrorDependentServicesRunning:   fmt.Errorf("ERROR_DEPENDENT_SERVICES_RUNNING"),
-	ErrorInvalidServiceControl:      fmt.Errorf("ERROR_INVALID_SERVICE_CONTROL"),
-	ErrorServiceRequestTimeout:      fmt.Errorf("Error service request timeout"),
-	ErrorServiceNoThread:            fmt.Errorf("A thread could not be created for the service."),
-	ErrorServiceDatabaseLocked:      fmt.Errorf("The service database is locked by the call to the BlockServiceDatabase method."),
-	ErrorServiceAlreadyRunning:      fmt.Errorf("Service already running!"),
-	ErrorInvalidServiceAccount:      fmt.Errorf("ERROR_INVALID_SERVICE_ACCOUNT"),
-	ErrorServiceDisabled:            fmt.Errorf("ERROR_SERVICE_DISABLED"),
-	ErrorCircularDependency:         fmt.Errorf("ERROR_CIRCULAR_DEPENDENCY"),
-	ErrorServiceDoesNotExist:        fmt.Errorf("ERROR_SERVICE_DOES_NOT_EXIST"),
-	ErrorServiceCannotAcceptControl: fmt.Errorf("ERROR_SERVICE_CANNOT_ACCEPT_CONTROL"),
-	ErrorServiceNotActive:           fmt.Errorf("ERROR_SERVICE_NOT_ACTIVE"),
-	ErrorServiceDependencyFail:      fmt.Errorf("The specified service depends on another service that has failed to start."),
-	ErrorServiceLogonFailed:         fmt.Errorf("The service did not start due to a logon failure."),
-	ErrorServiceMarkedForDelete:     fmt.Errorf("Service marked for delete."),
-	ErrorServiceDependencyDeleted:   fmt.Errorf("The specified service depends on a service that does not exist or has been marked for deletion."),
-	ErrorDuplicateServiceName:       fmt.Errorf("ERROR_DUPLICATE_SERVICE_NAME"),
-	ErrorShutdownInProgress:         fmt.Errorf("The system is shutting down."),
-}
-
-const (
-	StypeDisktree    uint32 = 0x00000000 // Disk drive
-	StypePrintq      uint32 = 0x00000001 // Print queue
-	StypeDevice      uint32 = 0x00000002 // Communication device
-	StypeIPC         uint32 = 0x00000003 // Interprocess communication (IPC)
-	StypeClusterFS   uint32 = 0x02000000 // A cluster share
-	StypeClusterSOFS uint32 = 0x04000000 // A Scale-Out cluster share
-	StypeClusterDFS  uint32 = 0x08000000 // A DFS share in a cluster
-	StypeSpecial     uint32 = 0x80000000 // Special share reserved for interprocess communication (IPC$) or remote administration of the server (ADMIN$). Can also refer to administrative shares such as C$, D$, E$, and so forth.
-	StypeTemporary   uint32 = 0x40000000 // A temporary share that is not persisted for creation each time the file server initializes.
-)
-
-var ShareTypeMap = map[uint32]string{
-	StypeDisktree:    "Disk Drive",
-	StypePrintq:      "Print Queue",
-	StypeDevice:      "Communication Device",
-	StypeIPC:         "IPC",
-	StypeClusterFS:   "Cluster Share",
-	StypeClusterSOFS: "Scale-Out cluster share",
-	StypeClusterDFS:  "DFS Share in cluster",
-	StypeSpecial:     "Hidden",
-	StypeTemporary:   "Temp",
-}
-
-// MS-SRVS Response codes from 2.2.2.10 Common Error Codes
-const (
-	SRVSErrorFileNotFound        uint32 = 2
-	SRVSErrorAccessDenied        uint32 = 5
-	SRVSErrorNotSupported        uint32 = 50
-	SRVSErrorDupName             uint32 = 52
-	SRVSErrorInvalidParameter    uint32 = 87
-	SRVSErrorInvalidLevel        uint32 = 124
-	SRVSErrorMoreData            uint32 = 234
-	SRVSErrorServiceDoesNotExist uint32 = 1060
-	SRVSErrorInvalidDomainName   uint32 = 1212
-	SRVSNERRUnknownDevDir        uint32 = 2116
-	SRVSNERRRedirectedPath       uint32 = 2117
-	SRVSNERRDuplicateShare       uint32 = 2118
-	SRVSNERRBufTooSmall          uint32 = 2123
-	SRVSNERRUserNotFound         uint32 = 2221
-	SRVSNERRNetNameNotFound      uint32 = 2310
-	SRVSNERRDeviceNotShared      uint32 = 2311
-	SRVSNERRClientNameNotFound   uint32 = 2312
-	SRVSNERRInvalidComputer      uint32 = 2351
-)
-
-var SRVSResponseCodeMap = map[uint32]error{
-	SRVSErrorFileNotFound:        fmt.Errorf("The system cannot find the file specified"),
-	SRVSErrorAccessDenied:        fmt.Errorf("The user does not have access to the requested information"),
-	SRVSErrorNotSupported:        fmt.Errorf("The server does not support branch cache"),
-	SRVSErrorDupName:             fmt.Errorf("A duplicate name exists on the network"),
-	SRVSErrorInvalidParameter:    fmt.Errorf("One or more of the specified parameters is invalid"),
-	SRVSErrorInvalidLevel:        fmt.Errorf("The value that is specified for the level parameter is invalid"),
-	SRVSErrorMoreData:            fmt.Errorf("More entries are available. Specify a large enough buffer to receive all entries"),
-	SRVSErrorServiceDoesNotExist: fmt.Errorf("The branch cache component does not exist as an installed service"),
-	SRVSErrorInvalidDomainName:   fmt.Errorf("The format of the specified NetBIOS name of a domain is invalid"),
-	SRVSNERRUnknownDevDir:        fmt.Errorf("The device or directory does not exist"),
-	SRVSNERRRedirectedPath:       fmt.Errorf("The operation is not valid for a redirected resource. The specified device name is assigned to a shared resource"),
-	SRVSNERRDuplicateShare:       fmt.Errorf("The share name is already in use on this server"),
-	SRVSNERRBufTooSmall:          fmt.Errorf("The client request succeeded. More entries are available. The buffer size that is specified by PreferedMaximumLength was too small to fit even a single entry"),
-	SRVSNERRUserNotFound:         fmt.Errorf("The user name could not be found"),
-	SRVSNERRNetNameNotFound:      fmt.Errorf("The share name does not exist"),
-	SRVSNERRDeviceNotShared:      fmt.Errorf("The device is not shared"),
-	SRVSNERRClientNameNotFound:   fmt.Errorf("A session does not exist with the computer name"),
-	SRVSNERRInvalidComputer:      fmt.Errorf("The computer name is not valid"),
-}
 
 // Unused
 //type RPCClient interface {
@@ -428,506 +174,6 @@ type UnicodeStr struct {
 	Padd          []byte `smb:"align:4"`
 }
 
-type Level struct {
-	Lvl uint32
-}
-
-type TotalEntries struct {
-	Padd []byte `smb:"align:4"`
-	Lvl  uint32
-}
-
-type ShareName struct {
-	UnicodeStr
-}
-
-type NetShareInfo1 struct {
-	ReferentId uint32
-	MaxCount   uint32
-}
-
-type ShareInfo1 struct {
-	Name    *UnicodeStr
-	Type    uint32
-	Comment *UnicodeStr
-}
-
-type NetShare struct {
-	Name    string
-	Comment string
-	Type    string
-	TypeId  uint32
-	Hidden  bool
-}
-
-type NetServerInfo100 struct {
-	platformId uint32
-	name       *UnicodeStr
-}
-
-type NetServerInfo101 struct {
-	NetServerInfo100
-	versionMajor uint32
-	versionMinor uint32
-	svType       uint32
-	comment      *UnicodeStr
-}
-
-type NetServerInfo102 struct {
-	NetServerInfo101
-	users    uint32
-	disc     int32
-	hidden   uint32
-	announce uint32
-	anndelta uint32
-	licences uint32
-	userpath *UnicodeStr
-}
-
-type NetServerInfo struct {
-	Level   uint32
-	Pointer interface{}
-}
-
-type ServerInfo100 struct {
-	PlatformId uint32
-	Name       string
-}
-
-type ServerInfo101 struct {
-	ServerInfo100
-	VersionMajor uint32
-	VersionMinor uint32
-	SvType       uint32
-	Comment      string
-}
-
-type ServerInfo102 struct {
-	ServerInfo101
-	Users    uint32
-	Disc     int32
-	Hidden   bool
-	Announce uint32
-	Anndelta uint32
-	Licences uint32
-	Userpath string
-}
-
-type SessionInfo0 struct {
-	Cname string
-}
-
-type SessionInfo1 struct {
-	SessionInfo0
-	Username  string
-	NumOpens  uint32
-	Time      uint32
-	IdleTime  uint32
-	UserFlags uint32 // Must be a combination of one or more of the values that are defined in 2.2.2.3
-}
-
-type SessionInfo2 struct {
-	SessionInfo1
-	ClType string
-}
-
-type SessionInfo10 struct {
-	SessionInfo0
-	Username string
-	Time     uint32
-	IdleTime uint32
-}
-
-type SessionInfo502 struct {
-	SessionInfo2
-	Transport string
-}
-
-// Could this be done in a better way? Perhaps with an interface?
-type NetSessionInfo0 struct {
-	Cname *UnicodeStr
-}
-
-type NetSessionInfo1 struct {
-	NetSessionInfo0
-	Username  *UnicodeStr
-	NumOpens  uint32
-	Time      uint32
-	IdleTime  uint32
-	UserFlags uint32 // Must be a combination of one or more of the values that are defined in 2.2.2.3
-}
-
-type NetSessionInfo2 struct {
-	NetSessionInfo1
-	ClType *UnicodeStr
-}
-
-type NetSessionInfo10 struct {
-	NetSessionInfo0
-	Username *UnicodeStr
-	Time     uint32
-	IdleTime uint32
-}
-
-type NetSessionInfo502 struct {
-	NetSessionInfo2
-	Transport *UnicodeStr
-}
-
-// Could this be done in a better way? Perhaps with an interface?
-type SessionInfoContainer struct {
-	EntriesRead uint32
-	Buffer      []interface{}
-}
-
-type SessionInfoContainer0 struct {
-	EntriesRead uint32
-	Buffer      []NetSessionInfo0
-}
-
-type SessionInfoContainer1 struct {
-	EntriesRead uint32
-	Buffer      []NetSessionInfo1
-}
-
-type SessionInfoContainer2 struct {
-	EntriesRead uint32
-	Buffer      []NetSessionInfo2
-}
-
-type SessionInfoContainer10 struct {
-	EntriesRead uint32
-	Buffer      []NetSessionInfo10
-}
-
-type SessionInfoContainer502 struct {
-	EntriesRead uint32
-	Buffer      []NetSessionInfo502
-}
-
-type ServiceConfig struct {
-	ServiceType      string
-	StartType        string
-	ErrorControl     string
-	BinaryPathName   string
-	LoadOrderGroup   string
-	TagId            uint32
-	Dependencies     string
-	ServiceStartName string
-	DisplayName      string
-}
-
-type NetShareCtr1 struct {
-	Count   uint32
-	Info    NetShareInfo1
-	Pointer []ShareInfo1
-}
-
-type NetShareCtr struct {
-	Ctr     uint32 // Type of Ctr struct in Pointer
-	Pointer interface{}
-}
-
-type ResumeHandle struct {
-	ReferentId uint32
-	Handle     uint32
-}
-
-type NetShareEnumAllRequest struct {
-	ServerName *UnicodeStr
-	Level
-	NetShareCtr
-	MaxBuffer uint32
-	ResumeHandle
-}
-
-type NetShareEnumAllResponse struct {
-	Level        uint32
-	NetShareCtr  *NetShareCtr
-	TotalEntries uint32
-	*ResumeHandle
-	WindowsError uint32
-}
-
-type NetServerGetInfoRequest struct {
-	ServerName *UnicodeStr
-	Level      uint32
-}
-
-type NetServerGetInfoResponse struct {
-	Info         *NetServerInfo
-	WindowsError uint32
-}
-
-type SessionEnum struct {
-	Level       uint32
-	SessionInfo interface{}
-}
-
-type NetSessionEnumRequest struct {
-	ServerName         *UnicodeStr
-	ClientName         *UnicodeStr
-	UserName           *UnicodeStr
-	Info               SessionEnum
-	PreferredMaxLength uint32
-	ResumeHandle       ResumeHandle
-}
-
-// Annoying to unmarshal because of NDR synax where lots and lots of ReferentId
-// Ptrs are part of the serialized data, and elements of the structs are NOT
-// serialized in order.
-type NetSessionEnumResponse struct {
-	Info         SessionEnum
-	TotalEntries uint32
-	ResumeHandle ResumeHandle
-	WindowsError uint32
-}
-
-type ROpenSCManagerWRequest struct {
-	MachineName   UnicodeStr
-	DatabaseName  UnicodeStr
-	DesiredAccess uint32
-}
-
-type ROpenSCManagerWResponse struct {
-	ContextHandle []byte `smb:"fixed:20"`
-	ReturnCode    uint32
-}
-
-type ROpenServiceWRequest struct {
-	SCContextHandle []byte `smb:"fixed:20"`
-	ServiceName     *UnicodeStr
-	DesiredAccess   uint32
-}
-
-type ROpenServiceWResponse struct {
-	ContextHandle []byte `smb:"fixed:20"`
-	ReturnCode    uint32
-}
-
-/*24 bytes between ErrorControl and BinaryPathName seems to contain
-  Maximum counts of 4 bytes for All the dimensions of the array.
-  That would be if the array contains 6 members (2 strings, 1 DWORD and then 3 more strings.
-  That would account for 6*4 bytes = 24. The values would be:
-  0x00020000, 0x00020004, 0x00000000, 0x00020008, 0x0002000c, 0x00020010
-  Not sure why this is so but the only Theory that seems to fit the data.*/
-
-// MS-SCMR Section 2.2.15
-type QueryServiceConfigW struct {
-	ServiceType   uint32
-	StartType     uint32
-	ErrorControl  uint32
-	ArrayMaxSizes []uint32 `smb:"count:6"` // The data after this is a complex structure
-	// and I simplify parsing here by ignoring the preceeding array max counts before each array
-	// item since this seems to be a Structure containing a Conformant and variying array.
-	BinaryPathName   UnicodeStr
-	LoadOrderGroup   UnicodeStr
-	TagId            uint32
-	Dependencies     UnicodeStr
-	ServiceStartName UnicodeStr
-	DisplayName      UnicodeStr
-}
-
-type RQueryServiceConfigWRequest struct {
-	ServiceHandle []byte `smb:"fixed:20"`
-	BufSize       uint32
-}
-
-type RQueryServiceConfigWResponse struct {
-	ServiceConfig *QueryServiceConfigW
-	BytesNeeded   uint32
-	ErrorCode     uint32
-}
-
-/*
-NOTE If A value is empty e.g. NULL, that is represented by 4 null bytes
-e.g., Don't specifying a BinaryPathName? Just put 0x00 0x00 0x00 0x00 as a null pointer
-*/
-type RChangeServiceConfigWRequest struct {
-	ServiceHandle    []byte `smb:"fixed:20"`
-	ServiceType      uint32
-	StartType        uint32
-	ErrorControl     uint32
-	BinaryPathName   *UnicodeStr
-	LoadOrderGroup   *UnicodeStr
-	TagId            uint32
-	Dependencies     *UnicodeStr // Not sure this is the correct type
-	DependSize       uint32
-	ServiceStartName *UnicodeStr
-	Password         *UnicodeStr // Not sure this is the correct type
-	PwSize           uint32
-	DisplayName      *UnicodeStr
-}
-
-type RChangeServiceConfigWResponse struct {
-	TagId      uint32
-	ReturnCode uint32
-}
-
-type RCreateServiceRequest struct {
-	SCContextHandle  []byte `smb:"fixed:20"`
-	ServiceName      *UnicodeStr
-	DisplayName      *UnicodeStr
-	DesiredAccess    uint32
-	ServiceType      uint32
-	StartType        uint32
-	ErrorControl     uint32
-	BinaryPathName   *UnicodeStr
-	LoadOrderGroup   *UnicodeStr
-	TagId            uint32
-	Dependencies     *UnicodeStr
-	DependSize       uint32
-	ServiceStartName *UnicodeStr
-	Password         *UnicodeStr
-	PwSize           uint32
-}
-
-type RCreateServiceResponse struct {
-	TagId         uint32
-	ContextHandle []byte `smb:"fixed:20"`
-	ReturnCode    uint32
-}
-
-type RQueryServiceStatusRequest struct {
-	ContextHandle []byte `smb:"fixed:20"`
-}
-
-type RQueryServiceStatusResponse struct {
-	ServiceStatus
-	ReturnCode uint32
-}
-
-type ServiceStatus struct {
-	ServiceType             uint32
-	CurrentState            uint32
-	ControlsAccepted        uint32
-	Win32ExitCode           uint32
-	ServiceSpecificExitCode uint32
-	CheckPoint              uint32
-	WaitHint                uint32
-}
-
-type RStartServiceWRequest struct {
-	ServiceHandle []byte `smb:"fixed:20"`
-	Argc          uint32 `smb:"count:Argv"`
-	Argv          []UnicodeStr
-}
-
-type RControlServiceRequest struct {
-	ServiceHandle []byte `smb:"fixed:20"`
-	Control       uint32
-}
-
-type RControlServiceResponse struct {
-	ServiceStatus
-	ReturnValue uint32
-}
-
-type RDeleteServiceRequest struct {
-	ServiceHandle []byte `smb:"fixed:20"`
-}
-
-type RDeleteServiceResponse struct {
-	ReturnCode uint32
-}
-
-type RCloseServiceHandleReq struct {
-	ServiceHandle []byte `smb:"fixed:20"`
-}
-
-type RCloseServiceHandleRes struct {
-	ContextHandle []byte `smb:"fixed:20"`
-	ReturnCode    uint32
-}
-
-func (s *RStartServiceWRequest) MarshalBinary(meta *encoder.Metadata) ([]byte, error) {
-	var ret []byte
-	w := bytes.NewBuffer(ret)
-
-	// Encode ServiceHandle
-	w.Write(s.ServiceHandle)
-	// Encode Argc
-	if err := binary.Write(w, binary.LittleEndian, s.Argc); err != nil {
-		return nil, err
-	}
-
-	// If Argc is 0, Argv will be a null pointer
-	if s.Argc == 0 {
-		w.Write([]byte{0x0, 0x0, 0x0, 0x0})
-		return w.Bytes(), nil
-	}
-
-	refId := make([]byte, 4)
-	_, err := rand.Read(refId)
-	if err != nil {
-		log.Errorln(err)
-		return nil, err
-	}
-
-	_, err = w.Write(refId)
-	if err != nil {
-		log.Errorln(err)
-		return nil, err
-	}
-
-	// Encode Max Element Count in array
-	if err := binary.Write(w, binary.LittleEndian, s.Argc); err != nil {
-		return nil, err
-	}
-
-	// Encode another RefId for each element in the array
-	// This is because according to NDR, the pointers are lifted outside the array element and into the parent structure
-	for i := 0; i < int(s.Argc); i++ {
-		_, err := rand.Read(refId)
-		if err != nil {
-			log.Errorln(err)
-			return nil, err
-		}
-
-		_, err = w.Write(refId)
-		if err != nil {
-			log.Errorln(err)
-			return nil, err
-		}
-	}
-
-	for i := 0; i < int(s.Argc); i++ {
-		// Encode MaxCount
-		if err := binary.Write(w, binary.LittleEndian, s.Argv[i].MaxCount); err != nil {
-			log.Errorln(err)
-			return nil, err
-		}
-		// Encode Offset
-		if err := binary.Write(w, binary.LittleEndian, s.Argv[i].Offset); err != nil {
-			log.Errorln(err)
-			return nil, err
-		}
-		// Encode ActualCount
-		if err := binary.Write(w, binary.LittleEndian, s.Argv[i].ActualCount); err != nil {
-			log.Errorln(err)
-			return nil, err
-		}
-		// Encode String
-		if err := binary.Write(w, binary.LittleEndian, s.Argv[i].EncodedString); err != nil {
-			log.Errorln(err)
-			return nil, err
-		}
-		// Encode Padding
-		if len(s.Argv[i].Padd) > 0 {
-			if err := binary.Write(w, binary.LittleEndian, s.Argv[i].Padd); err != nil {
-				log.Errorln(err)
-				return nil, err
-			}
-		}
-	}
-
-	return w.Bytes(), nil
-}
-
-func (self *RStartServiceWRequest) UnmarshalBinary(buf []byte, meta *encoder.Metadata) error {
-	return fmt.Errorf("NOT IMPLEMENTED UnmarshalBinary of RStartServiceWRequest")
-}
-
 func (s *ContextItems) MarshalBinary(meta *encoder.Metadata) ([]byte, error) {
 	//log.Debugln("In MarshalBinary for ContextItems")
 	var ret []byte
@@ -1026,258 +272,6 @@ func (self *ContextItem) UnmarshalBinary(buf []byte, meta *encoder.Metadata) err
 	self.BindMinVersion = binary.LittleEndian.Uint16(buf[22:24])
 	self.TransferUUID = buf[24:40]
 	self.TransferVersion = binary.LittleEndian.Uint32(buf[40:44])
-	return nil
-}
-
-func (self *RQueryServiceConfigWResponse) MarshalBinary(meta *encoder.Metadata) ([]byte, error) {
-	log.Errorln("NOT IMPLEMENTED MarshalBinary of RQueryServiceConfigWResponse")
-	return nil, nil
-}
-
-func (self *RQueryServiceConfigWResponse) UnmarshalBinary(buf []byte, meta *encoder.Metadata) error {
-	log.Debugln("In UnmarshalBinary for RQueryServiceconfigWResponse")
-	if len(buf) < 44 {
-		return fmt.Errorf("Buffer to small for RQueryServiceConfigWResponse")
-	}
-
-	conf := QueryServiceConfigW{
-		ServiceType:  binary.LittleEndian.Uint32(buf),
-		StartType:    binary.LittleEndian.Uint32(buf[4:]),
-		ErrorControl: binary.LittleEndian.Uint32(buf[8:]),
-	}
-
-	if len(buf) > 44 { // Perhaps full response
-		conf.ArrayMaxSizes = make([]uint32, 6)
-		conf.ArrayMaxSizes[0] = binary.LittleEndian.Uint32(buf[12:])
-		conf.ArrayMaxSizes[1] = binary.LittleEndian.Uint32(buf[16:])
-		conf.ArrayMaxSizes[2] = binary.LittleEndian.Uint32(buf[20:])
-		conf.ArrayMaxSizes[3] = binary.LittleEndian.Uint32(buf[24:])
-		conf.ArrayMaxSizes[4] = binary.LittleEndian.Uint32(buf[28:])
-		conf.ArrayMaxSizes[5] = binary.LittleEndian.Uint32(buf[32:])
-		//fmt.Printf("Array sizes: 0: %d, 1: %d, 2: %d, 3: %d, 4: %d, 5: %d\n", conf.ArrayMaxSizes[0], conf.ArrayMaxSizes[1], conf.ArrayMaxSizes[2], conf.ArrayMaxSizes[3], conf.ArrayMaxSizes[4], conf.ArrayMaxSizes[5])
-		offset := uint32(36)
-
-		us := UnicodeStr{
-			MaxCount:    binary.LittleEndian.Uint32(buf[offset:]),
-			Offset:      binary.LittleEndian.Uint32(buf[offset+4:]),
-			ActualCount: binary.LittleEndian.Uint32(buf[offset+8:]),
-		}
-		us.EncodedString = make([]byte, us.ActualCount*2)
-		copy(us.EncodedString, buf[offset+12:offset+12+us.ActualCount*2])
-		padd := (us.ActualCount * 2) % 4
-		if padd != 0 {
-			padd = 4 - padd
-			us.Padd = make([]byte, padd)
-		}
-		offset += 12 + us.ActualCount*2 + padd
-		conf.BinaryPathName = us
-
-		us = UnicodeStr{
-			MaxCount:    binary.LittleEndian.Uint32(buf[offset:]),
-			Offset:      binary.LittleEndian.Uint32(buf[offset+4:]),
-			ActualCount: binary.LittleEndian.Uint32(buf[offset+8:]),
-		}
-		us.EncodedString = make([]byte, us.ActualCount*2)
-		copy(us.EncodedString, buf[offset+12:offset+12+us.ActualCount*2])
-		padd = (us.ActualCount * 2) % 4
-		if padd != 0 {
-			padd = 4 - padd
-			us.Padd = make([]byte, padd)
-		}
-		offset += 12 + us.ActualCount*2 + padd
-		conf.LoadOrderGroup = us
-
-		if conf.ArrayMaxSizes[2] > 0 {
-			//TODO Fix this part if I ever see a response with a non-zero TagId
-			return fmt.Errorf("NOT IMPLEMENTED parsing of TagID from response when non-zero")
-
-			//fmt.Printf("Before tagId. Offset: %d\n", offset)
-			//tagMaxCount := binary.LittleEndian.Uint32(buf[offset:])
-			////tagOffset := binary.LittleEndian.Uint32(buf[offset+4:])
-			//tagActualCount := binary.LittleEndian.Uint32(buf[offset+8:])
-			//tagLength := 4*tagActualCount
-			//conf.TagId = binary.LittleEndian.Uint32(buf[offset+12:offset+12+tagLength])
-			//fmt.Printf("After tagid before offset bump. Offset: %d, tagActualCount: %d, tagMaxCount: %d\n", offset, tagActualCount, tagMaxCount)
-			//offset += 12+tagLength
-		}
-
-		//fmt.Printf("After tagId. Offset: %d, tagLength: %d\n", offset, tagLength)
-		// Dependencies
-		us = UnicodeStr{
-			MaxCount:    binary.LittleEndian.Uint32(buf[offset:]),
-			Offset:      binary.LittleEndian.Uint32(buf[offset+4:]),
-			ActualCount: binary.LittleEndian.Uint32(buf[offset+8:]),
-		}
-		us.EncodedString = make([]byte, us.ActualCount*2)
-		copy(us.EncodedString, buf[offset+12:offset+12+us.ActualCount*2])
-		padd = (us.ActualCount * 2) % 4
-		if padd != 0 {
-			padd = 4 - padd
-			us.Padd = make([]byte, padd)
-		}
-		offset += 12 + us.ActualCount*2 + padd
-		conf.Dependencies = us
-
-		// ServiceStartName
-		us = UnicodeStr{
-			MaxCount:    binary.LittleEndian.Uint32(buf[offset:]),
-			Offset:      binary.LittleEndian.Uint32(buf[offset+4:]),
-			ActualCount: binary.LittleEndian.Uint32(buf[offset+8:]),
-		}
-		us.EncodedString = make([]byte, us.ActualCount*2)
-		copy(us.EncodedString, buf[offset+12:offset+12+us.ActualCount*2])
-		padd = (us.ActualCount * 2) % 4
-		if padd != 0 {
-			padd = 4 - padd
-			us.Padd = make([]byte, padd)
-		}
-		offset += 12 + us.ActualCount*2 + padd
-		conf.ServiceStartName = us
-
-		// DisplayName
-		us = UnicodeStr{
-			MaxCount:    binary.LittleEndian.Uint32(buf[offset:]),
-			Offset:      binary.LittleEndian.Uint32(buf[offset+4:]),
-			ActualCount: binary.LittleEndian.Uint32(buf[offset+8:]),
-		}
-		us.EncodedString = make([]byte, us.ActualCount*2)
-		copy(us.EncodedString, buf[offset+12:offset+12+us.ActualCount*2])
-		padd = (us.ActualCount * 2) % 4
-		if padd != 0 {
-			padd = 4 - padd
-			us.Padd = make([]byte, padd)
-		}
-		offset += 12 + us.ActualCount*2 + padd
-		conf.DisplayName = us
-		self.BytesNeeded = binary.LittleEndian.Uint32(buf[len(buf)-8:])
-		self.ErrorCode = binary.LittleEndian.Uint32(buf[len(buf)-4:])
-	} else {
-		// Empty Arrays
-		self.BytesNeeded = binary.LittleEndian.Uint32(buf[36:])
-		self.ErrorCode = binary.LittleEndian.Uint32(buf[40:])
-	}
-	self.ServiceConfig = &conf
-	return nil
-}
-
-func (s *NetShareEnumAllRequest) MarshalBinary(meta *encoder.Metadata) ([]byte, error) {
-	log.Debugln("In MarshalBinary for NetShareEnumAllRequest")
-
-	var ret []byte
-	w := bytes.NewBuffer(ret)
-	serverName, err := encoder.Marshal(s.ServerName)
-	if err != nil {
-		return nil, err
-	}
-	w.Write(serverName)
-	binary.Write(w, binary.LittleEndian, s.Level.Lvl)
-	binary.Write(w, binary.LittleEndian, s.NetShareCtr.Ctr)
-	// TODO Do this properly e.g., check type of ctr and serialize it.
-	binary.Write(w, binary.LittleEndian, uint32(2))
-	binary.Write(w, binary.LittleEndian, make([]byte, 8))
-
-	binary.Write(w, binary.LittleEndian, s.MaxBuffer)
-	binary.Write(w, binary.LittleEndian, s.ResumeHandle.ReferentId)
-	binary.Write(w, binary.LittleEndian, s.ResumeHandle.Handle)
-
-	return w.Bytes(), nil
-}
-
-func (s *NetShareEnumAllRequest) UnmarshalBinary(buf []byte, meta *encoder.Metadata) error {
-	return fmt.Errorf("NOT IMPLEMENTED UnmarshalBinary of NetShareEnumAllRequest")
-}
-
-func (s *NetShareEnumAllResponse) MarshalBinary(meta *encoder.Metadata) ([]byte, error) {
-	return nil, fmt.Errorf("NOT IMPLEMENTED MarshaBinary of NetShareEnumAllResponse")
-}
-
-func parseNetShareCtr1(buf []byte) (*NetShareCtr1, error) {
-	log.Debugln("In parseNetShareCtr1")
-	offset := 0
-	res := &NetShareCtr1{
-		Count: binary.LittleEndian.Uint32(buf[offset+4:]), // Skip over ReferentId
-		Info: NetShareInfo1{
-			MaxCount: binary.LittleEndian.Uint32(buf[offset+12:]),
-		},
-	}
-	offset += 16
-	arr := make([]ShareInfo1, res.Count)
-	for i := 0; i < int(res.Count); i++ {
-		arr[i].Name = &UnicodeStr{
-			ReferentIdPtr: binary.LittleEndian.Uint32(buf[offset:]),
-		}
-		arr[i].Type = binary.LittleEndian.Uint32(buf[offset+4:])
-		arr[i].Comment = &UnicodeStr{
-			ReferentIdPtr: binary.LittleEndian.Uint32(buf[offset+8:]),
-		}
-		offset += 12
-	}
-	for i := 0; i < int(res.Count); i++ {
-		arr[i].Name.MaxCount = binary.LittleEndian.Uint32(buf[offset:])
-		arr[i].Name.Offset = binary.LittleEndian.Uint32(buf[offset+4:])
-		arr[i].Name.ActualCount = binary.LittleEndian.Uint32(buf[offset+8:])
-		nameLen := int(arr[i].Name.ActualCount)
-		arr[i].Name.EncodedString = buf[offset+12 : offset+12+nameLen*2]
-		paddLen := (nameLen * 2) % 4
-		if paddLen != 0 {
-			paddLen = 4 - paddLen
-		}
-		offset += 12 + nameLen*2 + paddLen
-
-		arr[i].Comment.MaxCount = binary.LittleEndian.Uint32(buf[offset:])
-		arr[i].Comment.Offset = binary.LittleEndian.Uint32(buf[offset+4:])
-		arr[i].Comment.ActualCount = binary.LittleEndian.Uint32(buf[offset+8:])
-		commentLen := int(arr[i].Comment.ActualCount)
-		arr[i].Comment.EncodedString = buf[offset+12 : offset+12+commentLen*2]
-		paddLen = (commentLen * 2) % 4
-		if paddLen != 0 {
-			paddLen = 4 - paddLen
-		}
-		offset += 12 + commentLen*2 + paddLen
-	}
-	res.Pointer = arr
-
-	return res, nil
-}
-
-func (s *NetShareEnumAllResponse) UnmarshalBinary(buf []byte, meta *encoder.Metadata) error {
-	log.Debugln("In UnmarshalBinary for NetShareEnumAllResponse")
-
-	res := NetShareEnumAllResponse{
-		Level: binary.LittleEndian.Uint32(buf),
-	}
-	if res.Level != 1 {
-		return fmt.Errorf("Unrecognized ShareInfo level %d\n", res.Level)
-	}
-	offset := 4
-
-	ctrType := binary.LittleEndian.Uint32(buf[offset:])
-	var ctrPtr interface{}
-	var err error
-	switch ctrType {
-	case 1:
-		offset += 4
-		ctrPtr, err = parseNetShareCtr1(buf[offset:])
-		if err != nil {
-			return err
-		}
-	default:
-		return fmt.Errorf("NOT IMPLEMENTED NetShareEnumAllResponse CtrType %d\n", ctrType)
-	}
-
-	res.NetShareCtr = &NetShareCtr{
-		Ctr:     ctrType,
-		Pointer: ctrPtr,
-	}
-
-	totalLen := len(buf)
-	res.TotalEntries = binary.LittleEndian.Uint32(buf[totalLen-16:])
-	res.ResumeHandle = &ResumeHandle{
-		Handle: binary.LittleEndian.Uint32(buf[totalLen-8:]),
-	}
-	res.WindowsError = binary.LittleEndian.Uint32(buf[totalLen-4:])
-
-	*s = res
 	return nil
 }
 
@@ -1413,72 +407,6 @@ func NewRequestReq(callId uint32, op uint16) (*RequestReq, error) {
 	}, nil
 }
 
-func NewNetShareEnumAllRequest(serverName string) *NetShareEnumAllRequest {
-	nr := NetShareEnumAllRequest{
-		ServerName: NewUnicodeStr(0x20000, serverName),
-		Level:      Level{Lvl: 1},
-		NetShareCtr: NetShareCtr{
-			Ctr:     1,
-			Pointer: make([]byte, 8),
-		},
-		MaxBuffer: 0xffffffff,
-		ResumeHandle: ResumeHandle{
-			ReferentId: 0x00020008,
-			Handle:     0,
-		},
-	}
-
-	return &nr
-}
-
-func NewNetServerGetInfoRequest(serverName string, level int) *NetServerGetInfoRequest {
-	if level < 100 || level > 102 {
-		log.Errorln("Invalid level for NetServerGetInfo request. Falling back to level 100")
-		level = 100
-	}
-	nr := NetServerGetInfoRequest{
-		Level: uint32(level),
-	}
-	if serverName != "" {
-		nr.ServerName = NewUnicodeStr(1, serverName)
-	}
-
-	return &nr
-}
-
-func NewNetSessionEnumRequest(clientName, userName string, level uint32) *NetSessionEnumRequest {
-	if (level > 2) && level != 10 && level != 502 {
-		// Valid levels are 0, 1, 2, 10, 502
-		log.Errorln("Invalid level for NetSessionEnum request. Falling back to level 10")
-		level = 10
-	}
-	nr := NetSessionEnumRequest{
-		Info: SessionEnum{Level: uint32(level)},
-	}
-	if clientName != "" {
-		nr.ClientName = NewUnicodeStr(1, clientName)
-	}
-	if userName != "" {
-		nr.UserName = NewUnicodeStr(2, userName)
-	}
-	nr.PreferredMaxLength = 0xffffffff
-	nr.ResumeHandle.ReferentId = 3
-
-	switch level {
-	case 0:
-		nr.Info.SessionInfo = SessionInfoContainer0{}
-	case 10:
-		nr.Info.SessionInfo = SessionInfoContainer10{}
-	case 502:
-		nr.Info.SessionInfo = SessionInfoContainer502{}
-	default:
-		log.Errorln("Not yet implemented level %d\n", level)
-		return nil
-	}
-
-	return &nr
-}
-
 func Bind(f *smb.File, interface_uuid string, majorVersion, minorVersion uint16, transfer_uuid string) (bind *ServiceBind, err error) {
 	log.Debugln("In Bind")
 	callId := rand.Uint32()
@@ -1528,828 +456,11 @@ func roundup(x, align int) int {
 	return (x + (align - 1)) &^ (align - 1)
 }
 
-func decodeServiceConfig(config *QueryServiceConfigW) (res ServiceConfig, err error) {
-	log.Debugln("In decodeServiceConfig")
-	if _, ok := ServiceTypeStatusMap[config.ServiceType]; !ok {
-		err = fmt.Errorf("Could not identify returned service type: %d\n", config.ServiceType)
-		fmt.Println(err)
-	}
-	res.ServiceType = ServiceTypeStatusMap[config.ServiceType]
-
-	if _, ok := StartTypeStatusMap[config.StartType]; !ok {
-		err = fmt.Errorf("Could not identify returned start type: %d\n", config.StartType)
-		fmt.Println(err)
-	}
-	res.StartType = StartTypeStatusMap[config.StartType]
-
-	if _, ok := ErrorControlStatusMap[config.ErrorControl]; !ok {
-		err = fmt.Errorf("Could not identify returned start type: %d\n", config.ErrorControl)
-		fmt.Println(err)
-	}
-	res.ErrorControl = ErrorControlStatusMap[config.ErrorControl]
-
-	res.BinaryPathName, err = encoder.FromUnicodeString(config.BinaryPathName.EncodedString)
-	if err != nil {
-		err = fmt.Errorf("Error decoding service config: %s\n", err)
-		fmt.Println(err)
-	}
-
-	res.LoadOrderGroup, err = encoder.FromUnicodeString(config.LoadOrderGroup.EncodedString)
-	if err != nil {
-		err = fmt.Errorf("Error decoding service config: %s\n", err)
-		fmt.Println(err)
-	}
-
-	res.TagId = config.TagId
-
-	res.Dependencies, err = encoder.FromUnicodeString(config.Dependencies.EncodedString)
-	if err != nil {
-		err = fmt.Errorf("Error decoding service config: %s\n", err)
-		fmt.Println(err)
-	}
-
-	res.ServiceStartName, err = encoder.FromUnicodeString(config.ServiceStartName.EncodedString)
-	if err != nil {
-		err = fmt.Errorf("Error decoding service config: %s\n", err)
-		fmt.Println(err)
-	}
-
-	res.DisplayName, err = encoder.FromUnicodeString(config.DisplayName.EncodedString)
-	if err != nil {
-		err = fmt.Errorf("Error decoding service config: %s\n", err)
-		fmt.Println(err)
-	}
-
-	return
-}
-
-func (sb *ServiceBind) NetShareEnumAll(host string) (res []NetShare, err error) {
-	log.Debugln("In NetShareEnumAll")
-	netReq := NewNetShareEnumAllRequest(host)
-	netBuf, err := encoder.Marshal(netReq)
-	if err != nil {
-		return
-	}
-
-	buffer, err := sb.MakeIoCtlRequest(SrvSvcOpNetShareEnumAll, netBuf)
-	if err != nil {
-		return
-	}
-
-	var response NetShareEnumAllResponse
-	err = encoder.Unmarshal(buffer, &response)
-	if err != nil {
-		return
-	}
-	res = make([]NetShare, response.TotalEntries)
-	var ctr1 *NetShareCtr1
-	ctr1 = response.NetShareCtr.Pointer.(*NetShareCtr1)
-	for i := 0; i < int(response.TotalEntries); i++ {
-		res[i].Name, err = encoder.FromUnicodeString(ctr1.Pointer[i].Name.EncodedString)
-		if err != nil {
-			return
-		}
-
-		res[i].Comment, err = encoder.FromUnicodeString(ctr1.Pointer[i].Comment.EncodedString)
-
-		// Parse the TYPE
-		t := ""
-		if (ctr1.Pointer[i].Type & StypeClusterDFS) == StypeClusterDFS {
-			t += ShareTypeMap[StypeClusterDFS]
-			res[i].TypeId = StypeClusterDFS
-		} else if (ctr1.Pointer[i].Type & StypeClusterSOFS) == StypeClusterSOFS {
-			t += ShareTypeMap[StypeClusterSOFS]
-			res[i].TypeId = StypeClusterSOFS
-		} else if (ctr1.Pointer[i].Type & StypeClusterFS) == StypeClusterFS {
-			t += ShareTypeMap[StypeClusterFS]
-			res[i].TypeId = StypeClusterFS
-		} else if (ctr1.Pointer[i].Type & StypeIPC) == StypeIPC {
-			t += ShareTypeMap[StypeIPC]
-			res[i].TypeId = StypeIPC
-		} else if (ctr1.Pointer[i].Type & StypeDevice) == StypeDevice {
-			t += ShareTypeMap[StypeDevice]
-			res[i].TypeId = StypeDevice
-		} else if (ctr1.Pointer[i].Type & StypePrintq) == StypePrintq {
-			t += ShareTypeMap[StypePrintq]
-			res[i].TypeId = StypePrintq
-		} else {
-			t += ShareTypeMap[StypeDisktree]
-			res[i].TypeId = StypeDisktree
-		}
-
-		if (ctr1.Pointer[i].Type & StypeSpecial) == StypeSpecial {
-			t += "_" + ShareTypeMap[StypeSpecial]
-			res[i].Hidden = true
-		} else if (ctr1.Pointer[i].Type & StypeTemporary) == StypeTemporary {
-			t += "_" + ShareTypeMap[StypeTemporary]
-		}
-		res[i].Type = t
-	}
-
-	return res, nil
-}
-
-/*
-Send a NetServerGetInfo request to the server. Level can be 100, 101, or 102
-*/
-func (sb *ServiceBind) NetServerGetInfo(host string, level int) (res *NetServerInfo, err error) {
-	log.Debugln("In NetServerGetInfo")
-	netReq := NewNetServerGetInfoRequest(host, level)
-	netBuf, err := netReq.MarshalBinary(nil)
-	if err != nil {
-		return
-	}
-
-	buffer, err := sb.MakeIoCtlRequest(SrvSvcOpNetServerGetInfo, netBuf)
-	if err != nil {
-		return
-	}
-
-	if len(buffer) < 12 {
-		return nil, fmt.Errorf("Server response to NetServerGetInfo was too small. Expected at atleast 12 bytes")
-	}
-	werror := binary.LittleEndian.Uint32(buffer[len(buffer)-4:])
-	if werror != 0 {
-		responseCode, found := SRVSResponseCodeMap[werror]
-		if !found {
-			err = fmt.Errorf("NetServerGetInfo returned unknown error code: %d\n", werror)
-			log.Errorln(err)
-			return
-		}
-		log.Debugf("NetServerGetInfo return error: %v\n", responseCode)
-		return nil, responseCode
-	}
-
-	var response NetServerGetInfoResponse
-	err = response.UnmarshalBinary(buffer, nil)
-	if err != nil {
-		return
-	}
-	switch response.Info.Level {
-	case 100:
-		ptr := response.Info.Pointer.(*NetServerInfo100)
-		si := ServerInfo100{
-			PlatformId: ptr.platformId,
-		}
-		si.Name, err = encoder.FromUnicodeString(ptr.name.EncodedString)
-		if err != nil {
-			log.Errorln(err)
-			return
-		}
-		res = &NetServerInfo{Level: 100, Pointer: &si}
-	case 101:
-		ptr := response.Info.Pointer.(*NetServerInfo101)
-		si := ServerInfo101{
-			ServerInfo100: ServerInfo100{PlatformId: ptr.platformId},
-			VersionMajor:  ptr.versionMajor,
-			VersionMinor:  ptr.versionMinor,
-			SvType:        ptr.svType,
-		}
-		si.Name, err = encoder.FromUnicodeString(ptr.name.EncodedString)
-		if err != nil {
-			log.Errorln(err)
-			return
-		}
-		si.Comment, err = encoder.FromUnicodeString(ptr.comment.EncodedString)
-		if err != nil {
-			log.Errorln(err)
-			return
-		}
-		res = &NetServerInfo{Level: 101, Pointer: &si}
-	case 102:
-		ptr := response.Info.Pointer.(*NetServerInfo102)
-		si := ServerInfo102{
-			ServerInfo101: ServerInfo101{
-				ServerInfo100: ServerInfo100{PlatformId: ptr.platformId},
-				VersionMajor:  ptr.versionMajor,
-				VersionMinor:  ptr.versionMinor,
-				SvType:        ptr.svType,
-			},
-			Users:    ptr.users,
-			Disc:     ptr.disc,
-			Hidden:   ptr.hidden > 0,
-			Announce: ptr.announce,
-			Anndelta: ptr.anndelta,
-			Licences: ptr.licences,
-		}
-		si.Name, err = encoder.FromUnicodeString(ptr.name.EncodedString)
-		if err != nil {
-			log.Errorln(err)
-			return
-		}
-		si.Comment, err = encoder.FromUnicodeString(ptr.comment.EncodedString)
-		if err != nil {
-			log.Errorln(err)
-			return
-		}
-		si.Userpath, err = encoder.FromUnicodeString(ptr.userpath.EncodedString)
-		if err != nil {
-			log.Errorln(err)
-			return
-		}
-		res = &NetServerInfo{Level: 102, Pointer: &si}
-	}
-
-	return
-}
-
-/*
-Send a NetSessionEnum request to the server. Level can be 0, 1, 2, 10 or 502
-But so far only level 0, 10 and 502 are implemented
-*/
-func (sb *ServiceBind) NetSessionEnum(clientName, username string, level int) (res *SessionEnum, err error) {
-	log.Debugln("In NetServerGetInfo")
-	if level < 0 {
-		return nil, fmt.Errorf("Only levels 0, 1, 2, 10 and 502 are valid")
-	}
-	netReq := NewNetSessionEnumRequest(clientName, username, uint32(level))
-	netBuf, err := netReq.MarshalBinary(nil)
-	if err != nil {
-		return
-	}
-
-	buffer, err := sb.MakeIoCtlRequest(SrvSvcOpNetrSessionEnum, netBuf)
-	if err != nil {
-		return
-	}
-
-	if len(buffer) < 20 {
-		return nil, fmt.Errorf("Server response to NetSessionEnum was too small. Expected at atleast 20 bytes")
-	}
-	werror := binary.LittleEndian.Uint32(buffer[len(buffer)-4:])
-	if werror != 0 {
-		responseCode, found := SRVSResponseCodeMap[werror]
-		if !found {
-			err = fmt.Errorf("NetServerGetInfo returned unknown error code: %d\n", werror)
-			log.Errorln(err)
-			return
-		}
-		log.Debugf("NetServerGetInfo return error: %v\n", responseCode)
-		return nil, responseCode
-	}
-
-	var response NetSessionEnumResponse
-	err = response.UnmarshalBinary(buffer, nil)
-	if err != nil {
-		return
-	}
-	res = &SessionEnum{}
-	switch response.Info.Level {
-	case 0:
-		res.Level = 0
-		ptr := response.Info.SessionInfo.(*SessionInfoContainer0)
-		sic := &SessionInfoContainer{EntriesRead: ptr.EntriesRead}
-		res.SessionInfo = sic
-		for i := 0; i < int(ptr.EntriesRead); i++ {
-			si := SessionInfo0{
-				Cname: "",
-			}
-			si.Cname, err = encoder.FromUnicodeString(ptr.Buffer[i].Cname.EncodedString)
-			if err != nil {
-				log.Errorln(err)
-				return
-			}
-			sic.Buffer = append(sic.Buffer, si)
-		}
-	case 10:
-		res.Level = 10
-		ptr := response.Info.SessionInfo.(*SessionInfoContainer10)
-		sic := &SessionInfoContainer{EntriesRead: ptr.EntriesRead}
-		res.SessionInfo = sic
-		for i := 0; i < int(ptr.EntriesRead); i++ {
-			si := SessionInfo10{
-				SessionInfo0: SessionInfo0{Cname: ""},
-				Username:     "",
-				Time:         ptr.Buffer[i].Time,
-				IdleTime:     ptr.Buffer[i].IdleTime,
-			}
-			si.Cname, err = encoder.FromUnicodeString(ptr.Buffer[i].Cname.EncodedString)
-			if err != nil {
-				log.Errorln(err)
-				return
-			}
-			si.Username, err = encoder.FromUnicodeString(ptr.Buffer[i].Username.EncodedString)
-			if err != nil {
-				log.Errorln(err)
-				return
-			}
-			sic.Buffer = append(sic.Buffer, si)
-		}
-	case 502:
-		res.Level = 502
-		ptr := response.Info.SessionInfo.(*SessionInfoContainer502)
-		sic := &SessionInfoContainer{EntriesRead: ptr.EntriesRead}
-		res.SessionInfo = sic
-		for i := 0; i < int(ptr.EntriesRead); i++ {
-			si := SessionInfo502{}
-			si.Cname, err = encoder.FromUnicodeString(ptr.Buffer[i].Cname.EncodedString)
-			if err != nil {
-				log.Errorln(err)
-				return
-			}
-			si.Username, err = encoder.FromUnicodeString(ptr.Buffer[i].Username.EncodedString)
-			if err != nil {
-				log.Errorln(err)
-				return
-			}
-			si.ClType, err = encoder.FromUnicodeString(ptr.Buffer[i].ClType.EncodedString)
-			if err != nil {
-				log.Errorln(err)
-				return
-			}
-			si.Transport, err = encoder.FromUnicodeString(ptr.Buffer[i].Transport.EncodedString)
-			if err != nil {
-				log.Errorln(err)
-				return
-			}
-			si.UserFlags = ptr.Buffer[i].UserFlags
-			si.NumOpens = ptr.Buffer[i].NumOpens
-			si.Time = ptr.Buffer[i].Time
-			si.IdleTime = ptr.Buffer[i].IdleTime
-			sic.Buffer = append(sic.Buffer, si)
-		}
-	default:
-		return nil, fmt.Errorf("Server returned response with info level %d which is not yet implementet\n", response.Info.Level)
-	}
-
-	return
-}
-
-func (sb *ServiceBind) openSCManager(desiredAccess uint32) (handle []byte, err error) {
-	log.Debugln("In openSCManager")
-	scReq := ROpenSCManagerWRequest{
-		MachineName:   *NewUnicodeStr(1, "DUMMY"),
-		DatabaseName:  *NewUnicodeStr(2, "ServicesActive"),
-		DesiredAccess: desiredAccess,
-	}
-	scBuf, err := encoder.Marshal(scReq)
-	if err != nil {
-		return
-	}
-
-	buffer, err := sb.MakeIoCtlRequest(SvcCtlROpenSCManagerW, scBuf)
-	if err != nil {
-		return
-	}
-
-	// Retrieve context handle from response
-	res := ROpenSCManagerWResponse{}
-	err = encoder.Unmarshal(buffer, &res)
-	if err != nil {
-		return
-	}
-
-	if res.ReturnCode != ErrorSuccess {
-		err = ServiceResponseCodeMap[res.ReturnCode]
-	}
-
-	handle = res.ContextHandle
-	return
-}
-
-func (sb *ServiceBind) openService(scHandle []byte, serviceName string, desiredAccess uint32) (handle []byte, err error) {
-	log.Debugln("In openService")
-	serviceReq := ROpenServiceWRequest{
-		SCContextHandle: scHandle,
-		ServiceName:     NewUnicodeStr(0, serviceName),
-		DesiredAccess:   desiredAccess,
-	}
-
-	serviceBuf, err := encoder.Marshal(serviceReq)
-	if err != nil {
-		return
-	}
-
-	buffer, err := sb.MakeIoCtlRequest(SvcCtlROpenServiceW, serviceBuf)
-	if err != nil {
-		return
-	}
-
-	// Retrieve context handle from response
-	res := ROpenServiceWResponse{}
-	err = encoder.Unmarshal(buffer, &res)
-	if err != nil {
-		return
-	}
-
-	if res.ReturnCode != ErrorSuccess {
-		err = ServiceResponseCodeMap[res.ReturnCode]
-	}
-
-	handle = res.ContextHandle
-	return
-}
-
-func (sb *ServiceBind) GetServiceStatus(serviceName string) (status uint32, err error) {
-	log.Debugln("In GetServiceStatus")
-	handle, err := sb.openSCManager(ServiceQueryStatus)
-	if err != nil {
-		return
-	}
-	defer sb.CloseServiceHandle(handle)
-
-	serviceHandle, err := sb.openService(handle, serviceName, ServiceQueryStatus)
-	if err != nil {
-		return
-	}
-	defer sb.CloseServiceHandle(serviceHandle)
-
-	ssReq := RQueryServiceStatusRequest{ContextHandle: serviceHandle}
-	ssBuf, err := encoder.Marshal(ssReq)
-	if err != nil {
-		return
-	}
-
-	buffer, err := sb.MakeIoCtlRequest(SvcCtlRQueryServiceStatus, ssBuf)
-	if err != nil {
-		return
-	}
-
-	// Retrieve context handle from response
-	res := RQueryServiceStatusResponse{}
-	err = encoder.Unmarshal(buffer, &res)
-	if err != nil {
-		return
-	}
-
-	if res.ReturnCode != ErrorSuccess {
-		err = ServiceResponseCodeMap[res.ReturnCode]
-		return
-	}
-
-	status = res.ServiceStatus.CurrentState
-	return
-}
-
-func (sb *ServiceBind) StartService(serviceName string, args []string) (err error) {
-	log.Debugln("In StartService")
-	handle, err := sb.openSCManager(ServiceStart)
-	if err != nil {
-		return
-	}
-	defer sb.CloseServiceHandle(handle)
-	serviceHandle, err := sb.openService(handle, serviceName, ServiceStart)
-	if err != nil {
-		return
-	}
-	defer sb.CloseServiceHandle(serviceHandle)
-
-	ssReq := RStartServiceWRequest{ServiceHandle: serviceHandle}
-	if args == nil || len(args) == 0 {
-		ssReq.Argc = 0
-		ssReq.Argv = nil
-	} else {
-		arr := make([]UnicodeStr, len(args))
-		refId := uint32(1)
-		ssReq.Argc = uint32(len(args))
-		for i := range args {
-			arr[i] = *NewUnicodeStr(refId, args[i])
-			refId++
-		}
-		ssReq.Argv = arr
-	}
-
-	ssBuf, err := encoder.Marshal(&ssReq)
-	if err != nil {
-		return
-	}
-
-	buffer, err := sb.MakeIoCtlRequest(SvcCtlRStartServiceW, ssBuf)
-	if err != nil {
-		return
-	}
-
-	// Retrieve context handle from response
-	returnValue := binary.LittleEndian.Uint32(buffer)
-	if returnValue != ErrorSuccess {
-		return ServiceResponseCodeMap[returnValue]
-	}
-
-	return
-}
-
-func (sb *ServiceBind) ControlService(serviceName string, control uint32) (err error) {
-	log.Debugln("In ControlService")
-	handle, err := sb.openSCManager(SCManagerConnect)
-	if err != nil {
-		return
-	}
-	defer sb.CloseServiceHandle(handle)
-	serviceHandle, err := sb.openService(handle, serviceName, ServicePauseContinue|ServiceInterrogate|ServiceStop)
-	if err != nil {
-		return
-	}
-	defer sb.CloseServiceHandle(serviceHandle)
-
-	csReq := RControlServiceRequest{
-		ServiceHandle: serviceHandle,
-		Control:       control,
-	}
-	csBuf, err := encoder.Marshal(csReq)
-	if err != nil {
-		return
-	}
-
-	buffer, err := sb.MakeIoCtlRequest(SvcCtlRControlService, csBuf)
-	if err != nil {
-		return
-	}
-
-	// Parse ServiceStatus
-	res := RControlServiceResponse{}
-	err = encoder.Unmarshal(buffer, &res)
-	if err != nil {
-		return
-	}
-
-	// Retrieve context handle from response
-	if res.ReturnValue != ErrorSuccess {
-		return ServiceResponseCodeMap[res.ReturnValue]
-	}
-
-	return
-}
-
-func (sb *ServiceBind) GetServiceConfig(serviceName string) (config ServiceConfig, err error) {
-	log.Debugln("In GetServiceConfig")
-	handle, err := sb.openSCManager(ServiceQueryConfig)
-	if err != nil {
-		return
-	}
-	defer sb.CloseServiceHandle(handle)
-	serviceHandle, err := sb.openService(handle, serviceName, ServiceQueryConfig)
-	if err != nil {
-		return
-	}
-	defer sb.CloseServiceHandle(serviceHandle)
-
-	innerReq := RQueryServiceConfigWRequest{
-		ServiceHandle: serviceHandle,
-		BufSize:       0,
-	}
-	innerBuf, err := encoder.Marshal(innerReq)
-	if err != nil {
-		return
-	}
-
-	// Make request to figure out buffer size
-	buffer, err := sb.MakeIoCtlRequest(SvcCtlRQueryServiceConfigW, innerBuf)
-	if err != nil {
-		return
-	}
-
-	// Parse response
-	res := RQueryServiceConfigWResponse{}
-	err = encoder.Unmarshal(buffer, &res)
-	if err != nil {
-		return
-	}
-
-	if res.ErrorCode != ErrorInsufficientBuffer {
-		err = ServiceResponseCodeMap[res.ErrorCode]
-		return
-	}
-
-	// Repeat request with allocated buffer size
-	innerReq.BufSize = res.BytesNeeded
-	innerBuf2, err := encoder.Marshal(innerReq)
-	if err != nil {
-		return
-	}
-
-	buffer, err = sb.MakeIoCtlRequest(SvcCtlRQueryServiceConfigW, innerBuf2)
-	if err != nil {
-		return
-	}
-
-	// Parse ServiceConfig
-	res = RQueryServiceConfigWResponse{}
-	err = encoder.Unmarshal(buffer, &res)
-	if err != nil {
-		return
-	}
-
-	if res.ErrorCode != ErrorSuccess {
-		err = ServiceResponseCodeMap[res.ErrorCode]
-		return
-	}
-
-	return decodeServiceConfig(res.ServiceConfig)
-}
-
-func (sb *ServiceBind) ChangeServiceConfig(
-	serviceName string,
-	serviceType, startType, errorControl uint32,
-	binaryPathName, serviceStartName, displayName string) (err error) {
-
-	log.Debugln("In ChangeServiceConfig")
-
-	handle, err := sb.openSCManager(ServiceChangeConfig)
-	if err != nil {
-		return
-	}
-	defer sb.CloseServiceHandle(handle)
-	serviceHandle, err := sb.openService(handle, serviceName, ServiceChangeConfig)
-	if err != nil {
-		return
-	}
-	defer sb.CloseServiceHandle(serviceHandle)
-
-	innerReq := RChangeServiceConfigWRequest{
-		ServiceHandle:  serviceHandle,
-		ServiceType:    serviceType,
-		StartType:      startType,
-		ErrorControl:   errorControl,
-		LoadOrderGroup: nil,
-		Dependencies:   nil,
-		DependSize:     0,
-		Password:       nil,
-		PwSize:         0,
-	}
-
-	if len(binaryPathName) > 0 {
-		innerReq.BinaryPathName = NewUnicodeStr(0x00000001, binaryPathName)
-	} else {
-		innerReq.BinaryPathName = nil
-	}
-
-	if len(serviceStartName) > 0 {
-		innerReq.ServiceStartName = NewUnicodeStr(0x00000002, serviceStartName)
-	} else {
-		innerReq.ServiceStartName = nil
-	}
-
-	if len(displayName) > 0 {
-		innerReq.DisplayName = NewUnicodeStr(0x00000003, displayName)
-	} else {
-		innerReq.DisplayName = nil
-	}
-	innerBuf, err := encoder.Marshal(innerReq)
-	if err != nil {
-		return
-	}
-
-	buffer, err := sb.MakeIoCtlRequest(SvcCtlRChangeServiceConfigW, innerBuf)
-	if err != nil {
-		return
-	}
-
-	// Parse ServiceConfig
-	res := RChangeServiceConfigWResponse{}
-	err = encoder.Unmarshal(buffer, &res)
-	if err != nil {
-		return
-	}
-	if res.ReturnCode != ErrorSuccess {
-		return ServiceResponseCodeMap[res.ReturnCode]
-	}
-
-	return
-}
-
-func (sb *ServiceBind) CreateService(
-	serviceName string,
-	serviceType, startType, errorControl uint32,
-	binaryPathName, serviceStartName, displayName string, startService bool) (err error) {
-
-	log.Debugln("In CreateService")
-
-	scHandle, err := sb.openSCManager(SCManagerCreateService)
-	if err != nil {
-		return
-	}
-	defer sb.CloseServiceHandle(scHandle)
-
-	innerReq := RCreateServiceRequest{
-		SCContextHandle:  scHandle,
-		ServiceName:      NewUnicodeStr(0, serviceName),
-		DisplayName:      NewUnicodeStr(1, displayName),
-		DesiredAccess:    ServiceAllAccess,
-		ServiceType:      serviceType,
-		StartType:        startType,
-		ErrorControl:     errorControl,
-		BinaryPathName:   NewUnicodeStr(0, binaryPathName),
-		LoadOrderGroup:   nil,
-		TagId:            0,
-		Dependencies:     nil,
-		DependSize:       0,
-		ServiceStartName: NewUnicodeStr(2, serviceStartName),
-	}
-
-	log.Debugf("ServiceName: %s\n", innerReq.ServiceName.EncodedString)
-	// To support specifying a password I must figure out how the encryption is
-	// performed as this parameter expects an encrypted passphrase with some
-	// session key
-	//if password != "" {
-	//    innerReq.Password = NewUnicodeStr(0, password)
-	//    innerReq.PwSize = uint32((len(password)+1) * 2) // Null-terminated unicode string
-	//}
-
-	innerBuf, err := encoder.Marshal(innerReq)
-	if err != nil {
-		return
-	}
-
-	buffer, err := sb.MakeIoCtlRequest(SvcCtlRCreateServiceW, innerBuf)
-	if err != nil {
-		return
-	}
-
-	// Parse ServiceConfig
-	res := RCreateServiceResponse{}
-	err = encoder.Unmarshal(buffer, &res)
-	if err != nil {
-		return
-	}
-	if res.ReturnCode != ErrorSuccess {
-		err = ServiceResponseCodeMap[res.ReturnCode]
-		return
-	}
-	defer sb.CloseServiceHandle(res.ContextHandle)
-
-	if startService {
-		ssReq := RStartServiceWRequest{ServiceHandle: res.ContextHandle}
-		ssReq.Argc = 0
-		// When Argc is 0 I need to marshal 0x00000000 for Argc and same for Argv e.g., 4 bytes combined of 0s
-
-		ssBuf, err2 := encoder.Marshal(&ssReq)
-		if err != nil {
-			return err2
-		}
-
-		buffer, err2 := sb.MakeIoCtlRequest(SvcCtlRStartServiceW, ssBuf)
-		if err != nil {
-			return err2
-		}
-
-		returnValue := binary.LittleEndian.Uint32(buffer)
-		if returnValue != ErrorSuccess {
-			return ServiceResponseCodeMap[returnValue]
-		}
-	}
-
-	return
-}
-
-func (sb *ServiceBind) DeleteService(serviceName string) (err error) {
-	scHandle, err := sb.openSCManager(SCManagerConnect)
-	if err != nil {
-		return
-	}
-	defer sb.CloseServiceHandle(scHandle)
-	handle, err := sb.openService(scHandle, serviceName, ServiceAllAccess)
-	if err != nil {
-		return
-	}
-	defer sb.CloseServiceHandle(handle)
-
-	// Attempt to stop the service before deletion
-	csReq := RControlServiceRequest{
-		ServiceHandle: handle,
-		Control:       ServiceControlStop,
-	}
-	csBuf, err := encoder.Marshal(csReq)
-	if err != nil {
-		return
-	}
-
-	_, err = sb.MakeIoCtlRequest(SvcCtlRControlService, csBuf)
-	if err != nil {
-		log.Errorln(err)
-		// Continue with deletion even if stop failed for some reason
-	}
-
-	innerReq := RDeleteServiceRequest{
-		ServiceHandle: handle,
-	}
-
-	innerBuf, err := encoder.Marshal(innerReq)
-	if err != nil {
-		return
-	}
-
-	buffer, err := sb.MakeIoCtlRequest(SvcCtlRDeleteService, innerBuf)
-	if err != nil {
-		return
-	}
-
-	// Parse ServiceConfig
-	res := RDeleteServiceResponse{}
-	err = encoder.Unmarshal(buffer, &res)
-	if err != nil {
-		return
-	}
-	if res.ReturnCode != ErrorSuccess {
-		err = ServiceResponseCodeMap[res.ReturnCode]
-		return
-	}
-	return
-}
-
 func (sb *ServiceBind) MakeIoCtlRequest(opcode uint16, innerBuf []byte) (result []byte, err error) {
 	callId := rand.Uint32()
 	req, err := NewRequestReq(callId, opcode)
 	if err != nil {
+		log.Errorln(err)
 		return
 	}
 
@@ -2362,11 +473,13 @@ func (sb *ServiceBind) MakeIoCtlRequest(opcode uint16, innerBuf []byte) (result 
 	// Encode DCERPC Request
 	buf, err := encoder.Marshal(req)
 	if err != nil {
+		log.Errorln(err)
 		return
 	}
 
 	ioCtlReq, err := sb.f.NewIoCTLReq(smb.FsctlPipeTransceive, buf)
 	if err != nil {
+		log.Errorln(err)
 		return
 	}
 
@@ -2384,48 +497,18 @@ func (sb *ServiceBind) MakeIoCtlRequest(opcode uint16, innerBuf []byte) (result 
 	var reqRes RequestRes
 	err = encoder.Unmarshal(ioCtlRes.Buffer, &reqRes)
 	if err != nil {
+		log.Errorln(err)
 		return
 	}
 
 	if reqRes.CallId != callId {
 		err = fmt.Errorf("Incorrect CallId on response. Sent %d and received %d\n", callId, reqRes.CallId)
+		log.Errorln(err)
 		return
 	}
 
 	// Return response data
 	return reqRes.Buffer, err
-}
-
-func (sb *ServiceBind) CloseServiceHandle(serviceHandle []byte) {
-	//log.Debugln("In CloseServiceHandle")
-	closeReq := RCloseServiceHandleReq{
-		ServiceHandle: serviceHandle,
-	}
-	closeBuf, err := encoder.Marshal(closeReq)
-	if err != nil {
-		fmt.Printf("Failed to encode close service handle request")
-		return
-	}
-
-	buffer, err := sb.MakeIoCtlRequest(SvcCtlRCloseServiceHandle, closeBuf)
-	if err != nil {
-		fmt.Printf("Failed to close service handle with error: %s\n", err)
-		return
-	}
-	res := RCloseServiceHandleRes{}
-	err = encoder.Unmarshal(buffer, &res)
-	if err != nil {
-		fmt.Printf("Failed to unmarshal response of close service handle")
-		return
-	}
-
-	// Retrieve context handle from response
-	//returnValue := binary.LittleEndian.Uint32(buffer)
-	if res.ReturnCode != ErrorSuccess {
-		log.Errorf("Failed to close service handle with error (return value: 0x%x): %v\n", res.ReturnCode, ServiceResponseCodeMap[res.ReturnCode])
-	}
-
-	return
 }
 
 func (s *UnicodeStr) MarshalBinary(meta *encoder.Metadata) ([]byte, error) {
@@ -2484,568 +567,5 @@ func (self *UnicodeStr) UnmarshalBinary(buf []byte, meta *encoder.Metadata) erro
 		self.Padd = make([]byte, paddLen)
 	}
 
-	return nil
-}
-
-func (self *NetServerGetInfoRequest) MarshalBinary(meta *encoder.Metadata) ([]byte, error) {
-	log.Debugln("In MarshalBinary for NetServerGetInfoRequest")
-
-	var ret []byte
-	w := bytes.NewBuffer(ret)
-	if self.ServerName != nil {
-		serverName, err := self.ServerName.MarshalBinary(nil)
-		if err != nil {
-			return nil, err
-		}
-		w.Write(serverName)
-	} else {
-		w.Write([]byte{0, 0, 0, 0})
-	}
-	binary.Write(w, binary.LittleEndian, self.Level)
-
-	return w.Bytes(), nil
-}
-
-func (self *NetServerGetInfoRequest) UnmarshalBinary(buf []byte, meta *encoder.Metadata) error {
-	return fmt.Errorf("NOT IMPLEMENTED UnmarshalBinary of NetServerGetInfoRequest")
-}
-
-func (self *NetServerGetInfoResponse) MarshalBinary(meta *encoder.Metadata) ([]byte, error) {
-	return nil, fmt.Errorf("NOT IMPLEMENTED MarshaBinary of NetServerGetInfoResponse")
-}
-
-func (self *NetServerGetInfoResponse) UnmarshalBinary(buf []byte, meta *encoder.Metadata) error {
-	log.Debugln("In UnmarshalBinary for NetServerGetInfoResponse")
-
-	/*
-		The serialized format is a bit confusing as the structs/data types are not just
-		serialized in order. Instead they are put in different places depending on what
-		comes before and what comes after.
-		In this case, the ReferentId Ptrs are lifted out of the UnicodeStr structs and the actual order is as seen below.
-		This is because according to NDR, the pointers are lifted outside the array element (strings) and into the parent structure
-	*/
-	offset := 0
-	self.Info = &NetServerInfo{}
-	self.Info.Level = binary.LittleEndian.Uint32(buf[offset:])
-	// Skip over a ReferentId Ptr of 4 bytes
-	offset += 8
-	switch self.Info.Level {
-	case 100:
-		if len(buf[offset:]) < 8 {
-			return fmt.Errorf("Buffer is too short to contain a NetServerInfo100 struct\n")
-		}
-		si := NetServerInfo100{}
-		si.platformId = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.name = &UnicodeStr{ReferentIdPtr: binary.LittleEndian.Uint32(buf[offset:])}
-		offset += 4
-
-		si.name.MaxCount = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.name.Offset = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.name.ActualCount = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.name.EncodedString = make([]byte, si.name.ActualCount*2)
-		copy(si.name.EncodedString, buf[offset+int(si.name.Offset):offset+int(si.name.Offset)+int(si.name.ActualCount)*2])
-		offset += int(si.name.ActualCount) * 2
-
-		paddLen := 4 - (offset % 4)
-		if paddLen != 4 {
-			offset += paddLen
-			si.name.Padd = make([]byte, paddLen)
-		}
-
-		self.Info.Pointer = &si
-	case 101:
-		if len(buf[offset:]) < 24 {
-			return fmt.Errorf("Buffer is too short to contain a NetServerInfo102 struct\n")
-		}
-		si := NetServerInfo101{}
-		si.platformId = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.name = &UnicodeStr{ReferentIdPtr: binary.LittleEndian.Uint32(buf[offset:])}
-		offset += 4
-		si.versionMajor = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.versionMinor = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.svType = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.comment = &UnicodeStr{ReferentIdPtr: binary.LittleEndian.Uint32(buf[offset:])}
-		offset += 4
-
-		si.name.MaxCount = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.name.Offset = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.name.ActualCount = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.name.EncodedString = make([]byte, si.name.ActualCount*2)
-		copy(si.name.EncodedString, buf[offset+int(si.name.Offset):offset+int(si.name.Offset)+int(si.name.ActualCount)*2])
-		offset += int(si.name.ActualCount) * 2
-
-		paddLen := 4 - (offset % 4)
-		if paddLen != 4 {
-			offset += paddLen
-			si.name.Padd = make([]byte, paddLen)
-		}
-
-		// Comment struct
-		si.comment.MaxCount = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.comment.Offset = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.comment.ActualCount = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.comment.EncodedString = make([]byte, int(si.comment.ActualCount*2))
-		copy(si.comment.EncodedString, buf[offset+int(si.comment.Offset):offset+int(si.comment.Offset)+int(si.comment.ActualCount)*2])
-		offset += int(si.comment.ActualCount) * 2
-		paddLen = 4 - (offset % 4)
-		if paddLen != 4 {
-			offset += paddLen
-			si.comment.Padd = make([]byte, paddLen)
-		}
-
-		self.Info.Pointer = &si
-	case 102:
-		/*
-			        Order of serialization:
-					platformId
-					ReferentId Ptr from name struct
-					versionMajor
-					versionMinor
-					svType
-					ReferentId Ptr from comment struct
-					Users
-					disc
-					hidden
-					announce
-					anndelta
-					licenses
-					ReferentId Ptr from UserPath struct
-
-					Then finally comes the content of the three UnicodeStr structs
-		*/
-		if len(buf[offset:]) < 52 {
-			return fmt.Errorf("Buffer is too short to contain a NetServerInfo102 struct\n")
-		}
-		si := NetServerInfo102{}
-		si.platformId = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.name = &UnicodeStr{ReferentIdPtr: binary.LittleEndian.Uint32(buf[offset:])}
-		offset += 4
-		si.versionMajor = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.versionMinor = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.svType = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.comment = &UnicodeStr{ReferentIdPtr: binary.LittleEndian.Uint32(buf[offset:])}
-		offset += 4
-		si.users = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.disc = int32(binary.LittleEndian.Uint32(buf[offset:]))
-		offset += 4
-		si.hidden = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.announce = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.anndelta = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.licences = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.userpath = &UnicodeStr{ReferentIdPtr: binary.LittleEndian.Uint32(buf[offset:])}
-		offset += 4
-
-		si.name.MaxCount = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.name.Offset = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.name.ActualCount = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		//offset = 72
-		si.name.EncodedString = make([]byte, si.name.ActualCount*2)
-		copy(si.name.EncodedString, buf[offset+int(si.name.Offset):offset+int(si.name.Offset)+int(si.name.ActualCount)*2])
-		offset += int(si.name.ActualCount) * 2
-
-		paddLen := 4 - (offset % 4)
-		if paddLen != 4 {
-			offset += paddLen
-			si.name.Padd = make([]byte, paddLen)
-		}
-
-		// Comment struct
-		si.comment.MaxCount = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.comment.Offset = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.comment.ActualCount = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.comment.EncodedString = make([]byte, int(si.comment.ActualCount*2))
-		copy(si.comment.EncodedString, buf[offset+int(si.comment.Offset):offset+int(si.comment.Offset)+int(si.comment.ActualCount)*2])
-		offset += int(si.comment.ActualCount) * 2
-		paddLen = 4 - (offset % 4)
-		if paddLen != 4 {
-			offset += paddLen
-			si.comment.Padd = make([]byte, paddLen)
-		}
-
-		// UserPath struct
-		si.userpath.MaxCount = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.userpath.Offset = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.userpath.ActualCount = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-		si.userpath.EncodedString = make([]byte, int(si.userpath.ActualCount*2))
-		copy(si.userpath.EncodedString, buf[offset+int(si.userpath.Offset):offset+int(si.userpath.Offset)+int(si.userpath.ActualCount)*2])
-		offset += int(si.userpath.ActualCount) * 2
-		paddLen = 4 - (offset % 4)
-		if paddLen != 4 {
-			offset += paddLen
-			si.userpath.Padd = make([]byte, paddLen)
-		}
-		self.Info.Pointer = &si
-	}
-
-	self.WindowsError = binary.LittleEndian.Uint32(buf[offset:])
-	return nil
-}
-
-func (self *NetSessionEnumRequest) MarshalBinary(meta *encoder.Metadata) ([]byte, error) {
-	log.Debugln("In MarshalBinary for NetSessionEnumRequest")
-
-	var ret []byte
-	w := bytes.NewBuffer(ret)
-	if self.ServerName != nil {
-		serverName, err := self.ServerName.MarshalBinary(nil)
-		if err != nil {
-			return nil, err
-		}
-		w.Write(serverName)
-	} else {
-		w.Write([]byte{0, 0, 0, 0})
-	}
-	if self.ClientName != nil {
-		clientName, err := self.ClientName.MarshalBinary(nil)
-		if err != nil {
-			return nil, err
-		}
-		w.Write(clientName)
-	} else {
-		w.Write([]byte{0, 0, 0, 0})
-	}
-
-	if self.UserName != nil {
-		userName, err := self.UserName.MarshalBinary(nil)
-		if err != nil {
-			return nil, err
-		}
-		w.Write(userName)
-	} else {
-		w.Write([]byte{0, 0, 0, 0})
-	}
-
-	// Encode the level
-	binary.Write(w, binary.LittleEndian, self.Info.Level)
-
-	switch self.Info.Level {
-	case 0:
-		refIdCtr := byte(1)
-		binary.Write(w, binary.LittleEndian, self.Info.Level)                 // Encode the level again
-		binary.Write(w, binary.LittleEndian, []byte{refIdCtr, 0x0, 0x0, 0x0}) // ReferentID
-		refIdCtr++
-
-		ptr := self.Info.SessionInfo.(SessionInfoContainer0)
-		binary.Write(w, binary.LittleEndian, ptr.EntriesRead)            // How many items in array (sessions)
-		binary.Write(w, binary.LittleEndian, []byte{0x0, 0x0, 0x0, 0x0}) // Null ptr
-		//TODO Add support for specifying an argument array?
-		if ptr.EntriesRead > 0 {
-			return nil, fmt.Errorf("Not yet implemented support for specifying NetSession0 array items")
-		}
-
-	case 10:
-		refIdCtr := byte(1)
-		binary.Write(w, binary.LittleEndian, self.Info.Level)                 // Encode the level again
-		binary.Write(w, binary.LittleEndian, []byte{refIdCtr, 0x0, 0x0, 0x0}) // ReferentID
-		refIdCtr++
-
-		ptr := self.Info.SessionInfo.(SessionInfoContainer10)
-		binary.Write(w, binary.LittleEndian, ptr.EntriesRead) // How many items in array (sessions)
-
-		if len(ptr.Buffer) > 0 {
-			// This block is probably unnecessary as the request will likely never contain any elements
-			binary.Write(w, binary.LittleEndian, []byte{refIdCtr, 0x0, 0x0, 0x0}) // ReferentID
-			refIdCtr++
-			binary.Write(w, binary.LittleEndian, ptr.EntriesRead) // Max Count
-			// Encode the ReferentID ptrs for each element in the array
-			for i := range ptr.Buffer {
-				if ptr.Buffer[i].Cname != nil {
-					binary.Write(w, binary.LittleEndian, ptr.Buffer[i].Cname.ReferentIdPtr)
-				} else {
-					// Not sure it is correct to encode a null pointer
-					binary.Write(w, binary.LittleEndian, []byte{0, 0, 0, 0})
-				}
-				if ptr.Buffer[i].Username != nil {
-					binary.Write(w, binary.LittleEndian, ptr.Buffer[i].Username.ReferentIdPtr)
-				} else {
-					// Not sure it is correct to encode a null pointer
-					binary.Write(w, binary.LittleEndian, []byte{0, 0, 0, 0})
-				}
-				// Encode the last members of the SessionInfo10 struct
-				binary.Write(w, binary.LittleEndian, ptr.Buffer[i].Time)
-				binary.Write(w, binary.LittleEndian, ptr.Buffer[i].IdleTime)
-			}
-			// Now encode the actual Client and UserNames if they are present
-			for i := range ptr.Buffer {
-				if ptr.Buffer[i].Cname != nil {
-					ptr.Buffer[i].Cname.ReferentIdPtr = 0 // Workaround to handle when the ReferentId is lifted out of the struct
-					cnameBuff, err := ptr.Buffer[i].Cname.MarshalBinary(nil)
-					if err != nil {
-						log.Errorln(err)
-						return nil, err
-					}
-					w.Write(cnameBuff)
-				}
-				if ptr.Buffer[i].Username != nil {
-					ptr.Buffer[i].Username.ReferentIdPtr = 0 // Workaround to handle when the ReferentId is lifted out of the struct
-					usernameBuff, err := ptr.Buffer[i].Username.MarshalBinary(nil)
-					if err != nil {
-						log.Errorln(err)
-						return nil, err
-					}
-					w.Write(usernameBuff)
-				}
-			}
-		} else {
-			binary.Write(w, binary.LittleEndian, []byte{0x0, 0x0, 0x0, 0x0}) // Null ptr
-		}
-	case 502:
-		refIdCtr := byte(1)
-		binary.Write(w, binary.LittleEndian, self.Info.Level)                 // Encode the level again
-		binary.Write(w, binary.LittleEndian, []byte{refIdCtr, 0x0, 0x0, 0x0}) // ReferentID
-		refIdCtr++
-
-		ptr := self.Info.SessionInfo.(SessionInfoContainer502)
-		binary.Write(w, binary.LittleEndian, ptr.EntriesRead)            // How many items in array (sessions)
-		binary.Write(w, binary.LittleEndian, []byte{0x0, 0x0, 0x0, 0x0}) // Null ptr
-		//TODO Add support for specifying an argument array?
-		if ptr.EntriesRead > 0 {
-			return nil, fmt.Errorf("Not yet implemented support for specifying NetSession502 array items")
-		}
-	default:
-		return nil, fmt.Errorf("Not implemented marshal of level %d\n", self.Info.Level)
-	}
-
-	binary.Write(w, binary.LittleEndian, self.PreferredMaxLength)
-	binary.Write(w, binary.LittleEndian, self.ResumeHandle.ReferentId)
-	binary.Write(w, binary.LittleEndian, self.ResumeHandle.Handle)
-
-	return w.Bytes(), nil
-}
-
-func (self *NetSessionEnumRequest) UnmarshalBinary(buf []byte, meta *encoder.Metadata) error {
-	return fmt.Errorf("NOT IMPLEMENTED UnmarshalBinary of NetSessionEnumRequest")
-}
-
-func (self *NetSessionEnumResponse) MarshalBinary(meta *encoder.Metadata) ([]byte, error) {
-	return nil, fmt.Errorf("NOT IMPLEMENTED MarshaBinary of NetSessionEnumResponse")
-}
-
-func (self *NetSessionEnumResponse) UnmarshalBinary(buf []byte, meta *encoder.Metadata) error {
-	log.Debugln("In UnmarshalBinary for NetSessionEnumResponse")
-
-	offset := 0
-	self.Info.Level = binary.LittleEndian.Uint32(buf[offset:])
-	offset += 8 // Level is encoded twice because it is the discriminator for the SessionEnum Union
-
-	switch self.Info.Level {
-	case 0:
-		offset += 4 // Skip referrent ID
-		container := SessionInfoContainer0{}
-		self.Info.SessionInfo = &container
-		container.EntriesRead = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-
-		offset += 4 // Skip referrent ID
-		if container.EntriesRead > 0 {
-			offset += 4 // Skip Max count in front of the array
-			container.Buffer = make([]NetSessionInfo0, container.EntriesRead)
-			for i := 0; i < int(container.EntriesRead); i++ {
-				offset += 4 // Skip ReferentID ptrs for Cname
-			}
-			for i := 0; i < int(container.EntriesRead); i++ {
-				// Decode the Cname
-				container.Buffer[i].Cname = &UnicodeStr{
-					MaxCount:    binary.LittleEndian.Uint32(buf[offset:]),
-					Offset:      binary.LittleEndian.Uint32(buf[offset+4:]),
-					ActualCount: binary.LittleEndian.Uint32(buf[offset+8:]),
-				}
-				offset += 12
-				container.Buffer[i].Cname.EncodedString = make([]byte, container.Buffer[i].Cname.ActualCount*2)
-				copy(container.Buffer[i].Cname.EncodedString, buf[offset+int(container.Buffer[i].Cname.Offset):offset+int(container.Buffer[i].Cname.Offset)+int(container.Buffer[i].Cname.ActualCount*2)])
-				offset += int(container.Buffer[i].Cname.ActualCount * 2)
-				paddLen := 4 - (offset % 4)
-				if paddLen != 4 {
-					offset += paddLen
-					container.Buffer[i].Cname.Padd = make([]byte, paddLen)
-				}
-			}
-		}
-	case 10:
-		offset += 4 // Skip referrent ID
-		container := SessionInfoContainer10{}
-		self.Info.SessionInfo = &container
-		container.EntriesRead = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-
-		offset += 4 // Skip referrent ID
-		if container.EntriesRead > 0 {
-			offset += 4 // Skip Max count in front of the array
-			container.Buffer = make([]NetSessionInfo10, container.EntriesRead)
-			for i := 0; i < int(container.EntriesRead); i++ {
-				offset += 8 // Skip ReferentID ptrs for Cname and Username
-				container.Buffer[i].Time = binary.LittleEndian.Uint32(buf[offset:])
-				offset += 4
-				container.Buffer[i].IdleTime = binary.LittleEndian.Uint32(buf[offset:])
-				offset += 4
-			}
-			for i := 0; i < int(container.EntriesRead); i++ {
-				// Decode the Cname
-				//fmt.Printf("Offset before Cname Max Count: %d\n", offset)
-				container.Buffer[i].Cname = &UnicodeStr{
-					MaxCount:    binary.LittleEndian.Uint32(buf[offset:]),
-					Offset:      binary.LittleEndian.Uint32(buf[offset+4:]),
-					ActualCount: binary.LittleEndian.Uint32(buf[offset+8:]),
-				}
-				offset += 12
-				container.Buffer[i].Cname.EncodedString = make([]byte, container.Buffer[i].Cname.ActualCount*2)
-				copy(container.Buffer[i].Cname.EncodedString, buf[offset+int(container.Buffer[i].Cname.Offset):offset+int(container.Buffer[i].Cname.Offset)+int(container.Buffer[i].Cname.ActualCount*2)])
-				offset += int(container.Buffer[i].Cname.ActualCount * 2)
-				paddLen := 4 - (offset % 4)
-				if paddLen != 4 {
-					offset += paddLen
-					container.Buffer[i].Cname.Padd = make([]byte, paddLen)
-				}
-
-				// Decode the Username
-				container.Buffer[i].Username = &UnicodeStr{
-					MaxCount:    binary.LittleEndian.Uint32(buf[offset:]),
-					Offset:      binary.LittleEndian.Uint32(buf[offset+4:]),
-					ActualCount: binary.LittleEndian.Uint32(buf[offset+8:]),
-				}
-				offset += 12
-				container.Buffer[i].Username.EncodedString = make([]byte, container.Buffer[i].Username.ActualCount*2)
-				copy(container.Buffer[i].Username.EncodedString, buf[offset+int(container.Buffer[i].Username.Offset):offset+int(container.Buffer[i].Username.Offset)+int(container.Buffer[i].Username.ActualCount*2)])
-				offset += int(container.Buffer[i].Username.ActualCount * 2)
-				paddLen = 4 - (offset % 4)
-				if paddLen != 4 {
-					offset += paddLen
-					container.Buffer[i].Username.Padd = make([]byte, paddLen)
-				}
-			}
-		}
-	case 502:
-		offset += 4 // Skip referrent ID
-		container := SessionInfoContainer502{}
-		self.Info.SessionInfo = &container
-		container.EntriesRead = binary.LittleEndian.Uint32(buf[offset:])
-		offset += 4
-
-		offset += 4 // Skip referrent ID
-		if container.EntriesRead > 0 {
-			offset += 4 // Skip Max count in front of the array
-			container.Buffer = make([]NetSessionInfo502, container.EntriesRead)
-			for i := 0; i < int(container.EntriesRead); i++ {
-				offset += 8 // Skip ReferentID ptrs for Cname and Username
-				container.Buffer[i].NumOpens = binary.LittleEndian.Uint32(buf[offset:])
-				offset += 4
-				container.Buffer[i].Time = binary.LittleEndian.Uint32(buf[offset:])
-				offset += 4
-				container.Buffer[i].IdleTime = binary.LittleEndian.Uint32(buf[offset:])
-				offset += 4
-				container.Buffer[i].UserFlags = binary.LittleEndian.Uint32(buf[offset:])
-				offset += 4
-				offset += 8 // Skip ReferentID ptrs for ClientType and Transport
-			}
-			for i := 0; i < int(container.EntriesRead); i++ {
-				// Decode the Cname
-				container.Buffer[i].Cname = &UnicodeStr{
-					MaxCount:    binary.LittleEndian.Uint32(buf[offset:]),
-					Offset:      binary.LittleEndian.Uint32(buf[offset+4:]),
-					ActualCount: binary.LittleEndian.Uint32(buf[offset+8:]),
-				}
-				offset += 12
-				container.Buffer[i].Cname.EncodedString = make([]byte, container.Buffer[i].Cname.ActualCount*2)
-				copy(container.Buffer[i].Cname.EncodedString, buf[offset+int(container.Buffer[i].Cname.Offset):offset+int(container.Buffer[i].Cname.Offset)+int(container.Buffer[i].Cname.ActualCount*2)])
-				offset += int(container.Buffer[i].Cname.ActualCount * 2)
-				paddLen := 4 - (offset % 4)
-				if paddLen != 4 {
-					offset += paddLen
-					container.Buffer[i].Cname.Padd = make([]byte, paddLen)
-				}
-
-				// Decode the Username
-				container.Buffer[i].Username = &UnicodeStr{
-					MaxCount:    binary.LittleEndian.Uint32(buf[offset:]),
-					Offset:      binary.LittleEndian.Uint32(buf[offset+4:]),
-					ActualCount: binary.LittleEndian.Uint32(buf[offset+8:]),
-				}
-				offset += 12
-				container.Buffer[i].Username.EncodedString = make([]byte, container.Buffer[i].Username.ActualCount*2)
-				copy(container.Buffer[i].Username.EncodedString, buf[offset+int(container.Buffer[i].Username.Offset):offset+int(container.Buffer[i].Username.Offset)+int(container.Buffer[i].Username.ActualCount*2)])
-				offset += int(container.Buffer[i].Username.ActualCount * 2)
-				paddLen = 4 - (offset % 4)
-				if paddLen != 4 {
-					offset += paddLen
-					container.Buffer[i].Username.Padd = make([]byte, paddLen)
-				}
-
-				// Decode the ClientType
-				container.Buffer[i].ClType = &UnicodeStr{
-					MaxCount:    binary.LittleEndian.Uint32(buf[offset:]),
-					Offset:      binary.LittleEndian.Uint32(buf[offset+4:]),
-					ActualCount: binary.LittleEndian.Uint32(buf[offset+8:]),
-				}
-				offset += 12
-				container.Buffer[i].ClType.EncodedString = make([]byte, container.Buffer[i].ClType.ActualCount*2)
-				copy(container.Buffer[i].ClType.EncodedString, buf[offset+int(container.Buffer[i].ClType.Offset):offset+int(container.Buffer[i].ClType.Offset)+int(container.Buffer[i].ClType.ActualCount*2)])
-				offset += int(container.Buffer[i].ClType.ActualCount * 2)
-				paddLen = 4 - (offset % 4)
-				if paddLen != 4 {
-					offset += paddLen
-					container.Buffer[i].ClType.Padd = make([]byte, paddLen)
-				}
-
-				// Decode the Transport
-				container.Buffer[i].Transport = &UnicodeStr{
-					MaxCount:    binary.LittleEndian.Uint32(buf[offset:]),
-					Offset:      binary.LittleEndian.Uint32(buf[offset+4:]),
-					ActualCount: binary.LittleEndian.Uint32(buf[offset+8:]),
-				}
-				offset += 12
-				container.Buffer[i].Transport.EncodedString = make([]byte, container.Buffer[i].Transport.ActualCount*2)
-				copy(container.Buffer[i].Transport.EncodedString, buf[offset+int(container.Buffer[i].Transport.Offset):offset+int(container.Buffer[i].Transport.Offset)+int(container.Buffer[i].Transport.ActualCount*2)])
-				offset += int(container.Buffer[i].Transport.ActualCount * 2)
-				paddLen = 4 - (offset % 4)
-				if paddLen != 4 {
-					offset += paddLen
-					container.Buffer[i].Transport.Padd = make([]byte, paddLen)
-				}
-			}
-		}
-
-	default:
-		return fmt.Errorf("Not implemented UnmarshalBinary for level %d\n", self.Info.Level)
-	}
-
-	self.TotalEntries = binary.LittleEndian.Uint32(buf[offset:])
-	offset += 4
-	self.ResumeHandle.ReferentId = binary.LittleEndian.Uint32(buf[offset:])
-	offset += 4
-	self.ResumeHandle.Handle = binary.LittleEndian.Uint32(buf[offset:])
-	offset += 4
-	self.WindowsError = binary.LittleEndian.Uint32(buf[offset:])
 	return nil
 }
