+++ smb/dcerpc/scmr.go
@@ -1,0 +1,1668 @@
+// MIT License
+//
+// # Copyright (c) 2023 Jimmy Fjällid
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+//
+// The marshal/unmarshal of requerst and responses according to the NDR syntax
+// has been implemented on a per RPC request basis and not in any complete way.
+// As such, for each new functionality, a manual marshal and unmarshal method
+// has to be written for the relevant messages. This makes it a bit cumbersome
+// to implement new features, so at some point a major rewrite should be
+// performed to ideally handle the NDR syntax dynamically.
+
+package dcerpc
+
+import (
+	"bytes"
+	"encoding/binary"
+	"fmt"
+	"io"
+	"math/rand"
+
+	"github.com/jfjallid/go-smb/smb/encoder"
+)
+
+// MS-SCMR Operations OP Codes
+const (
+	SvcCtlRCloseServiceHandle   uint16 = 0
+	SvcCtlRControlService       uint16 = 1
+	SvcCtlRDeleteService        uint16 = 2
+	SvcCtlRQueryServiceStatus   uint16 = 6
+	SvcCtlRChangeServiceConfigW uint16 = 11
+	SvcCtlRCreateServiceW       uint16 = 12
+	SvcCtlROpenSCManagerW       uint16 = 15
+	SvcCtlROpenServiceW         uint16 = 16
+	SvcCtlRQueryServiceConfigW  uint16 = 17
+	SvcCtlRStartServiceW        uint16 = 19
+)
+
+// MS-SCMR (svcctl) Section 2.2.15 ServiceType merged with Section 2.2.47 dwServiceType
+const (
+	ServiceKernelDriver       uint32 = 0x00000001
+	ServiceFileSystemDriver   uint32 = 0x00000002
+	ServiceWin32OwnProcess    uint32 = 0x00000010
+	ServiceWin32ShareProcess  uint32 = 0x00000020
+	ServiceInteractiveProcess uint32 = 0x00000100
+)
+
+var ServiceTypeStatusMap = map[uint32]string{
+	ServiceKernelDriver:       "SERVICE_KERNEL_DRIVER",
+	ServiceFileSystemDriver:   "SERVICE_FILE_SYSTEM_DRIVER",
+	ServiceWin32OwnProcess:    "SERIVCE_WIN32_OWN_PROCESS",
+	ServiceWin32ShareProcess:  "SERVICE_WIN32_SHARE_PROCESS",
+	ServiceInteractiveProcess: "SERVICE_INTERACTIVE_PROCESS",
+}
+
+// MS-SCMR (svcctl) Table 2.2.15 StartType
+const (
+	ServiceBootStart   uint32 = 0x00000000
+	ServiceSystemStart uint32 = 0x00000001
+	ServiceAutoStart   uint32 = 0x00000002
+	ServiceDemandStart uint32 = 0x00000003
+	ServiceDisabled    uint32 = 0x00000004
+)
+
+var StartTypeStatusMap = map[uint32]string{
+	ServiceBootStart:   "SERVICE_BOOT_START",
+	ServiceSystemStart: "SERVICE_SYSTEM_START",
+	ServiceAutoStart:   "SERIVCE_AUTO_START",
+	ServiceDemandStart: "SERVICE_DEMAND_START",
+	ServiceDisabled:    "SERVICE_DISABLED",
+}
+
+// MS-SCMR (svcctl) Table 2.2.15 ErrorControl
+const (
+	ServiceErrorIgnore   uint32 = 0x00000000
+	ServiceErrorNormal   uint32 = 0x00000001
+	ServiceErrorSevere   uint32 = 0x00000002
+	ServiceErrorCritical uint32 = 0x00000003
+)
+
+var ErrorControlStatusMap = map[uint32]string{
+	ServiceErrorIgnore:   "SERVICE_ERROR_IGNORE",
+	ServiceErrorNormal:   "SERVICE_ERROR_NORMAL",
+	ServiceErrorSevere:   "SERIVCE_ERROR_SEVERE",
+	ServiceErrorCritical: "SERVICE_ERROR_CRITICAL",
+}
+
+// MS-SCMR (svcctl) Table 3.1.4
+const (
+	ServiceAllAccess           uint32 = 0x000F01FF //In addition to all access rights in this table, SERVICE_ALL_ACCESS includes Delete (DE), Read Control (RC), Write DACL (WD), and Write Owner (WO) access, as specified in ACCESS_MASK (section 2.4.3) of [MS-DTYP].
+	ServiceChangeConfig        uint32 = 0x00000002 //Required to change the configuration of a service.
+	ServiceEnumerateDependents uint32 = 0x00000008 //Required to enumerate the services installed on the server.
+	ServiceInterrogate         uint32 = 0x00000080 //Required to request immediate status from the service.
+	ServicePauseContinue       uint32 = 0x00000040 //Required to pause or continue the service.
+	ServiceQueryConfig         uint32 = 0x00000001 //Required to query the service configuration.
+	ServiceQueryStatus         uint32 = 0x00000004 //Required to request the service status.
+	ServiceStart               uint32 = 0x00000010 //Required to start the service.
+	ServiceStop                uint32 = 0x00000020 //Required to stop the service
+	ServiceUserDefinedControl  uint32 = 0x00000100 //Required to specify a user-defined control code.
+	ServiceSetStatus           uint32 = 0x00008000 //Required for a service to set its status.
+
+	SCManagerLock             uint32 = 0x00000008 //Required to lock the SCM database.
+	SCManagerCreateService    uint32 = 0x00000002 //Required for a service to be created.
+	SCManagerEnumerateService uint32 = 0x00000004 //Required to enumerate a service.
+	SCManagerConnect          uint32 = 0x00000001 //Required to connect to the SCM.
+	SCManagerQueryLockStatus  uint32 = 0x00000010 //Required to query the lock status of the SCM database.
+	SCManagerModifyBootConfig uint32 = 0x0020     //Required to call the RNotifyBootConfigStatus method.
+)
+
+// MS-SCMR Section 2.2.47 dwCurrentState
+const (
+	ServiceContinuePending uint32 = 0x00000005
+	ServicePausePending    uint32 = 0x00000006
+	ServicePaused          uint32 = 0x00000007
+	ServiceRunning         uint32 = 0x00000004
+	ServiceStartPending    uint32 = 0x00000002
+	ServiceStopPending     uint32 = 0x00000003
+	ServiceStopped         uint32 = 0x00000001
+)
+
+var ServiceStatusMap = map[uint32]string{
+	ServiceContinuePending: "SERVICE_CONTINUE_PENDING",
+	ServicePausePending:    "SERVICE_PAUSE_PENDING",
+	ServicePaused:          "SERVICE_PAUSED",
+	ServiceRunning:         "SERVICE_RUNNING",
+	ServiceStartPending:    "SERVICE_START_PENDING",
+	ServiceStopPending:     "SERVICE_STOP_PENDING",
+	ServiceStopped:         "SERVICE_STOPPED",
+}
+
+// MS-SCMR Section 3.1.4.2 dwControl
+const (
+	ServiceControlContinue       uint32 = 0x00000003
+	ServiceControlInterrogate    uint32 = 0x00000004
+	ServiceControlNetbindadd     uint32 = 0x00000007
+	ServiceControlNetbinddisable uint32 = 0x0000000A
+	ServiceControlNetbindenable  uint32 = 0x00000009
+	ServiceControlNetbindremove  uint32 = 0x00000008
+	ServiceControlParamChange    uint32 = 0x00000006
+	ServiceControlPause          uint32 = 0x00000002
+	ServiceControlStop           uint32 = 0x00000001
+)
+
+// MS-SCMR Section 3.1.4.11 RChangeServiceConfigW
+const (
+	ServiceNoChange uint32 = 0xffffffff
+)
+
+// MS-SCMR Response codes from multiple sections: 3.1.4.2, 3.1.4.11, 3.1.4.17, 3.1.4.19
+const (
+	ErrorSuccess                    uint32 = 0  // Successfully started the service
+	ErrorFileNotFound               uint32 = 2  // The system cannot find the file specified.
+	ErrorPathNotFound               uint32 = 3  // The system cannot find the path specified.
+	ErrorAccessDenied               uint32 = 5  // The SERVICE_START access right had not been granted to the caller when the RPC context handle to the service record was created.
+	ErrorInvalidHandle              uint32 = 6  // The handle is no longer valid.
+	ErrorInvalidParameter           uint32 = 87 // A parameter that was specified is invalid.
+	ErrorInsufficientBuffer         uint32 = 122
+	ErrorDependentServicesRunning   uint32 = 1051
+	ErrorInvalidServiceControl      uint32 = 1052
+	ErrorServiceRequestTimeout      uint32 = 1053 // The process for the service was started, but it did not respond within an implementation-specific time-out.
+	ErrorServiceNoThread            uint32 = 1054 // A thread could not be created for the service.
+	ErrorServiceDatabaseLocked      uint32 = 1055 // The service database is locked by the call to the BlockServiceDatabase method.
+	ErrorServiceAlreadyRunning      uint32 = 1056 // The ServiceStatus.dwCurrentState in the service record is not set to SERVICE_STOPPED.
+	ErrorInvalidServiceAccount      uint32 = 1057 // The user account name specified in the lpServiceStartName parameter does not exist.
+	ErrorServiceDisabled            uint32 = 1058 // The service cannot be started because the Start field in the service record is set to SERVICE_DISABLED.
+	ErrorCircularDependency         uint32 = 1059 // A circular dependency was specified.
+	ErrorServiceDoesNotExist        uint32 = 1060 // The service record with a specified display name does not exist in the SCM database
+	ErrorServiceCannotAcceptControl uint32 = 1061
+	ErrorServiceNotActive           uint32 = 1062
+	ErrorServiceDependencyFail      uint32 = 1068 // The specified service depends on another service that has failed to start.
+	ErrorServiceLogonFailed         uint32 = 1069 // The service did not start due to a logon failure.
+	ErrorServiceMarkedForDelete     uint32 = 1072 // The RDeleteService method has been called for the service record identified by the hService parameter.
+	ErrorServiceDependencyDeleted   uint32 = 1075 // The specified service depends on a service that does not exist or has been marked for deletion.
+	ErrorDuplicateServiceName       uint32 = 1078 // The lpDisplayName matches either the ServiceName or the DisplayName of another service record in the service control manager database.
+	ErrorShutdownInProgress         uint32 = 1115 // The system is shutting down.
+)
+
+var ServiceResponseCodeMap = map[uint32]error{
+	ErrorSuccess:                    fmt.Errorf("Successfully started the service"),
+	ErrorFileNotFound:               fmt.Errorf("The system cannot find the file specified."),
+	ErrorPathNotFound:               fmt.Errorf("The system cannot find the path specified."),
+	ErrorAccessDenied:               fmt.Errorf("ERROR_ACCESS_DENIED"),
+	ErrorInvalidHandle:              fmt.Errorf("The handle is no longer valid."),
+	ErrorInvalidParameter:           fmt.Errorf("A parameter that was specified is invalid."),
+	ErrorInsufficientBuffer:         fmt.Errorf("ERROR_INSUFFICIENT_BUFFER"),
+	ErrorDependentServicesRunning:   fmt.Errorf("ERROR_DEPENDENT_SERVICES_RUNNING"),
+	ErrorInvalidServiceControl:      fmt.Errorf("ERROR_INVALID_SERVICE_CONTROL"),
+	ErrorServiceRequestTimeout:      fmt.Errorf("Error service request timeout"),
+	ErrorServiceNoThread:            fmt.Errorf("A thread could not be created for the service."),
+	ErrorServiceDatabaseLocked:      fmt.Errorf("The service database is locked by the call to the BlockServiceDatabase method."),
+	ErrorServiceAlreadyRunning:      fmt.Errorf("Service already running!"),
+	ErrorInvalidServiceAccount:      fmt.Errorf("ERROR_INVALID_SERVICE_ACCOUNT"),
+	ErrorServiceDisabled:            fmt.Errorf("ERROR_SERVICE_DISABLED"),
+	ErrorCircularDependency:         fmt.Errorf("ERROR_CIRCULAR_DEPENDENCY"),
+	ErrorServiceDoesNotExist:        fmt.Errorf("ERROR_SERVICE_DOES_NOT_EXIST"),
+	ErrorServiceCannotAcceptControl: fmt.Errorf("ERROR_SERVICE_CANNOT_ACCEPT_CONTROL"),
+	ErrorServiceNotActive:           fmt.Errorf("ERROR_SERVICE_NOT_ACTIVE"),
+	ErrorServiceDependencyFail:      fmt.Errorf("The specified service depends on another service that has failed to start."),
+	ErrorServiceLogonFailed:         fmt.Errorf("The service did not start due to a logon failure."),
+	ErrorServiceMarkedForDelete:     fmt.Errorf("Service marked for delete."),
+	ErrorServiceDependencyDeleted:   fmt.Errorf("The specified service depends on a service that does not exist or has been marked for deletion."),
+	ErrorDuplicateServiceName:       fmt.Errorf("ERROR_DUPLICATE_SERVICE_NAME"),
+	ErrorShutdownInProgress:         fmt.Errorf("The system is shutting down."),
+}
+
+type ServiceConfig struct {
+	ServiceType      string
+	StartType        string
+	ErrorControl     string
+	BinaryPathName   string
+	LoadOrderGroup   string
+	TagId            uint32
+	Dependencies     string
+	ServiceStartName string
+	DisplayName      string
+}
+
+type ResumeHandle struct {
+	ReferentId uint32
+	Handle     uint32
+}
+
+/*
+DWORD ROpenSCManagerW(
+
+	[in, string, unique, range(0, SC_MAX_COMPUTER_NAME_LENGTH)]
+	SVCCTL_HANDLEW lpMachineName,
+	[in, string, unique, range(0, SC_MAX_NAME_LENGTH)]
+	wchar_t* lpDatabaseName,
+	[in] DWORD dwDesiredAccess,
+	[out] LPSC_RPC_HANDLE lpScHandle
+
+);
+*/
+type ROpenSCManagerWRequest struct {
+	MachineName   string
+	DatabaseName  string
+	DesiredAccess uint32
+}
+
+type ROpenSCManagerWResponse struct {
+	ContextHandle [20]byte
+	ReturnCode    uint32
+}
+
+/*
+DWORD ROpenServiceW(
+
+	[in] SC_RPC_HANDLE hSCManager,
+	[in, string, range(0, SC_MAX_NAME_LENGTH)]
+	wchar_t* lpServiceName,
+	[in] DWORD dwDesiredAccess,
+	[out] LPSC_RPC_HANDLE lpServiceHandle
+
+);
+*/
+type ROpenServiceWRequest struct {
+	SCContextHandle []byte
+	ServiceName     string
+	DesiredAccess   uint32
+}
+
+type ROpenServiceWResponse struct {
+	ContextHandle []byte
+	ReturnCode    uint32
+}
+
+type RCloseServiceHandleReq struct {
+	ServiceHandle []byte
+}
+
+type RCloseServiceHandleRes struct {
+	ContextHandle []byte
+	ReturnCode    uint32
+}
+
+type RQueryServiceStatusReq struct {
+	ContextHandle []byte
+}
+
+/*
+	typedef struct {
+	    DWORD dwServiceType;
+	    DWORD dwCurrentState;
+	    DWORD dwControlsAccepted;
+	    DWORD dwWin32ExitCode;
+	    DWORD dwServiceSpecificExitCode;
+	    DWORD dwCheckPoint;
+	    DWORD dwWaitHint;
+	} SERVICE_STATUS,
+
+*LPSERVICE_STATUS;
+*/
+type ServiceStatus struct {
+	ServiceType             uint32
+	CurrentState            uint32
+	ControlsAccepted        uint32
+	Win32ExitCode           uint32
+	ServiceSpecificExitCode uint32
+	CheckPoint              uint32
+	WaitHint                uint32
+}
+
+/*
+DWORD RQueryServiceStatus(
+
+	[in] SC_RPC_HANDLE hService,
+	[out] LPSERVICE_STATUS lpServiceStatus
+
+);
+*/
+type RQueryServiceStatusRes struct {
+	ServiceStatus *ServiceStatus
+	ReturnCode    uint32
+}
+
+// MS-SCMR Section 2.2.15
+/*
+typedef struct _QUERY_SERVICE_CONFIGW {
+DWORD dwServiceType;
+DWORD dwStartType;
+DWORD dwErrorControl;
+[string,range(0, 8 * 1024)] LPWSTR lpBinaryPathName;
+[string,range(0, 8 * 1024)] LPWSTR lpLoadOrderGroup;
+DWORD dwTagId;
+[string,range(0, 8 * 1024)] LPWSTR lpDependencies;
+[string,range(0, 8 * 1024)] LPWSTR lpServiceStartName;
+[string,range(0, 8 * 1024)] LPWSTR lpDisplayName;
+} QUERY_SERVICE_CONFIGW,
+*LPQUERY_SERVICE_CONFIGW;
+*/
+type QueryServiceConfigW struct {
+	ServiceType      uint32
+	StartType        uint32
+	ErrorControl     uint32
+	BinaryPathName   string
+	LoadOrderGroup   string
+	TagId            uint32
+	Dependencies     string // Array?
+	ServiceStartName string
+	DisplayName      string
+}
+
+/*
+DWORD RQueryServiceConfigW(
+
+	[in] SC_RPC_HANDLE hService,
+	[out] LPQUERY_SERVICE_CONFIGW lpServiceConfig,
+	[in, range(0, 1024*8)] DWORD cbBufSize,
+	[out] LPBOUNDED_DWORD_8K pcbBytesNeeded
+
+);
+*/
+type RQueryServiceConfigWReq struct {
+	ServiceHandle []byte
+	BufSize       uint32
+}
+
+type RQueryServiceConfigWRes struct {
+	ServiceConfig *QueryServiceConfigW
+	BytesNeeded   uint32
+	ErrorCode     uint32
+}
+
+/*
+DWORD RChangeServiceConfigW(
+
+	[in] SC_RPC_HANDLE hService,
+	[in] DWORD dwServiceType,
+	[in] DWORD dwStartType,
+	[in] DWORD dwErrorControl,
+	[in, string, unique, range(0, SC_MAX_PATH_LENGTH)]
+	    wchar_t* lpBinaryPathName,
+	[in, string, unique, range(0, SC_MAX_NAME_LENGTH)]
+	    wchar_t* lpLoadOrderGroup,
+	[in, out, unique] LPDWORD lpdwTagId,
+	[in, unique, size_is(dwDependSize)]
+	    LPBYTE lpDependencies,
+	[in, range(0, SC_MAX_DEPEND_SIZE)]
+	    DWORD dwDependSize,
+	[in, string, unique, range(0, SC_MAX_ACCOUNT_NAME_LENGTH)]
+	    wchar_t* lpServiceStartName,
+	[in, unique, size_is(dwPwSize)]
+	    LPBYTE lpPassword,
+	[in, range(0, SC_MAX_PWD_SIZE)]
+	    DWORD dwPwSize,
+	[in, string, unique, range(0, SC_MAX_NAME_LENGTH)]
+	    wchar_t* lpDisplayName
+
+);
+*/
+type RChangeServiceConfigWReq struct {
+	ServiceHandle    []byte
+	ServiceType      uint32
+	StartType        uint32
+	ErrorControl     uint32
+	BinaryPathName   string
+	LoadOrderGroup   string
+	TagId            uint32
+	Dependencies     string
+	DependSize       uint32
+	ServiceStartName string
+	// RPC over SMB requires password encryption with session key
+	// So have to encrypt the password before calling the marshal function
+	Password    []byte // []byte instead of string to support encryption
+	PwSize      uint32
+	DisplayName string
+}
+
+type RChangeServiceConfigWRes struct {
+	TagId      uint32
+	ReturnCode uint32
+}
+
+type RCreateServiceRequest struct {
+	SCContextHandle  []byte `smb:"fixed:20"`
+	ServiceName      *UnicodeStr
+	DisplayName      *UnicodeStr
+	DesiredAccess    uint32
+	ServiceType      uint32
+	StartType        uint32
+	ErrorControl     uint32
+	BinaryPathName   *UnicodeStr
+	LoadOrderGroup   *UnicodeStr
+	TagId            uint32
+	Dependencies     *UnicodeStr
+	DependSize       uint32
+	ServiceStartName *UnicodeStr
+	Password         *UnicodeStr
+	PwSize           uint32
+}
+
+type RCreateServiceResponse struct {
+	TagId         uint32
+	ContextHandle []byte `smb:"fixed:20"`
+	ReturnCode    uint32
+}
+
+type RStartServiceWRequest struct {
+	ServiceHandle []byte `smb:"fixed:20"`
+	Argc          uint32 `smb:"count:Argv"`
+	Argv          []UnicodeStr
+}
+
+type RControlServiceRequest struct {
+	ServiceHandle []byte `smb:"fixed:20"`
+	Control       uint32
+}
+
+type RControlServiceResponse struct {
+	ServiceStatus
+	ReturnValue uint32
+}
+
+type RDeleteServiceRequest struct {
+	ServiceHandle []byte `smb:"fixed:20"`
+}
+
+type RDeleteServiceResponse struct {
+	ReturnCode uint32
+}
+
+func (self *ROpenSCManagerWRequest) MarshalBinary() (res []byte, err error) {
+	log.Debugln("In MarshalBinary for ROpenSCManagerWRequest")
+
+	var ret []byte
+	w := bytes.NewBuffer(ret)
+	refId := uint32(1)
+
+	// Pointer to a conformant and varying string, so include ReferentId Ptr and MaxCount
+	_, err = writeConformantVaryingStringPtr(w, self.MachineName, refId)
+	if err != nil {
+		log.Errorln(err)
+		return nil, err
+	}
+	if self.MachineName != "" {
+		refId++
+	}
+
+	// Pointer to a conformant and varying string, so include ReferentId Ptr and MaxCount
+	_, err = writeConformantVaryingStringPtr(w, self.DatabaseName, refId)
+	if err != nil {
+		log.Errorln(err)
+		return nil, err
+	}
+	if self.DatabaseName != "" {
+		refId++
+	}
+
+	err = binary.Write(w, le, self.DesiredAccess)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	return w.Bytes(), nil
+}
+
+func (self *ROpenSCManagerWRequest) UnmarshalBinary(buf []byte) error {
+	return fmt.Errorf("NOT IMPLEMENTED UnmarshalBinary of ROpenSCManagerWRequest")
+}
+
+func (self *ROpenSCManagerWResponse) MarshalBinary() ([]byte, error) {
+	return nil, fmt.Errorf("NOT IMPLEMENTED MarshalBinary of ROpenSCManagerWRequest")
+}
+
+func (self *ROpenSCManagerWResponse) UnmarshalBinary(buf []byte) (err error) {
+	log.Debugln("In UnmarshalBinary for ROpenSCManagerWResponse")
+
+	r := bytes.NewReader(buf)
+	err = binary.Read(r, le, &self.ContextHandle)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+	err = binary.Read(r, le, &self.ReturnCode)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	return
+}
+
+func (self *ROpenServiceWRequest) MarshalBinary() (res []byte, err error) {
+	log.Debugln("In MarshalBinary for ROpenServiceWRequest")
+
+	var ret []byte
+	w := bytes.NewBuffer(ret)
+
+	if len(self.SCContextHandle) != 20 {
+		return nil, fmt.Errorf("Invalid size of SCContextHandle!")
+	}
+
+	_, err = w.Write(self.SCContextHandle[:20])
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+	// Pointer to a conformant and varying string, so include MaxCount
+	// Not sure why there is no ReferentId here as the IDL specifies it as a pointer
+	_, err = writeConformantVaryingString(w, self.ServiceName)
+	if err != nil {
+		log.Errorln(err)
+		return nil, err
+	}
+
+	err = binary.Write(w, le, self.DesiredAccess)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	return w.Bytes(), nil
+}
+
+func (self *ROpenServiceWRequest) UnmarshalBinary(buf []byte) error {
+	return fmt.Errorf("NOT IMPLEMENTED UnmarshalBinary of ROpenServiceWRequest")
+}
+
+func (self *ROpenServiceWResponse) MarshalBinary() ([]byte, error) {
+	return nil, fmt.Errorf("NOT IMPLEMENTED MarshalBinary of ROpenServiceWRequest")
+}
+
+func (self *ROpenServiceWResponse) UnmarshalBinary(buf []byte) (err error) {
+	log.Debugln("In UnmarshalBinary for ROpenServiceWResponse")
+
+	self.ContextHandle = make([]byte, 20)
+	r := bytes.NewReader(buf)
+	err = binary.Read(r, le, &self.ContextHandle)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+	err = binary.Read(r, le, &self.ReturnCode)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	return
+}
+
+func (self *RCloseServiceHandleReq) MarshalBinary() (res []byte, err error) {
+	log.Debugln("In MarshalBinary for RCloseServiceHandleReq")
+
+	var ret []byte
+	w := bytes.NewBuffer(ret)
+
+	if len(self.ServiceHandle) != 20 {
+		return nil, fmt.Errorf("Invalid size of ServiceHandle!")
+	}
+
+	_, err = w.Write(self.ServiceHandle[:20])
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	return w.Bytes(), nil
+}
+
+func (self *RCloseServiceHandleReq) UnmarshalBinary(buf []byte) error {
+	return fmt.Errorf("NOT IMPLEMENTED UnmarshalBinary of RCloseServiceHandleReq")
+}
+
+func (self *RCloseServiceHandleRes) MarshalBinary() ([]byte, error) {
+	return nil, fmt.Errorf("NOT IMPLEMENTED MarshalBinary of RCloseServiceHandleRes")
+}
+
+func (self *RCloseServiceHandleRes) UnmarshalBinary(buf []byte) (err error) {
+	log.Debugln("In UnmarshalBinary for RCloseServiceHandleRes")
+
+	self.ContextHandle = make([]byte, 20)
+	r := bytes.NewReader(buf)
+	err = binary.Read(r, le, &self.ContextHandle)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+	err = binary.Read(r, le, &self.ReturnCode)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	return
+}
+
+func (self *RQueryServiceStatusReq) MarshalBinary() (res []byte, err error) {
+	log.Debugln("In MarshalBinary for RQueryServiceStatusReq")
+
+	var ret []byte
+	w := bytes.NewBuffer(ret)
+
+	if len(self.ContextHandle) != 20 {
+		return nil, fmt.Errorf("Invalid size of ContextHandle!")
+	}
+
+	_, err = w.Write(self.ContextHandle[:20])
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	return w.Bytes(), nil
+}
+
+func (self *RQueryServiceStatusReq) UnmarshalBinary(buf []byte) error {
+	return fmt.Errorf("NOT IMPLEMENTED UnmarshalBinary of RQueryServiceStatusReq")
+}
+
+func (self *RQueryServiceStatusRes) MarshalBinary() ([]byte, error) {
+	return nil, fmt.Errorf("NOT IMPLEMENTED MarshalBinary of RQueryServiceStatusRes")
+}
+
+func (self *RQueryServiceStatusRes) UnmarshalBinary(buf []byte) (err error) {
+	log.Debugln("In UnmarshalBinary for RQueryServiceStatusRes")
+	r := bytes.NewReader(buf)
+
+	// Not sure why there is no RefId Ptr for the ServiceStatus struct ptr
+
+	self.ServiceStatus = &ServiceStatus{}
+	err = binary.Read(r, le, &self.ServiceStatus.ServiceType)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+	err = binary.Read(r, le, &self.ServiceStatus.CurrentState)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+	err = binary.Read(r, le, &self.ServiceStatus.ControlsAccepted)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+	err = binary.Read(r, le, &self.ServiceStatus.Win32ExitCode)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+	err = binary.Read(r, le, &self.ServiceStatus.ServiceSpecificExitCode)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+	err = binary.Read(r, le, &self.ServiceStatus.CheckPoint)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+	err = binary.Read(r, le, &self.ServiceStatus.WaitHint)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+	err = binary.Read(r, le, &self.ReturnCode)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	return
+}
+
+func (self *RQueryServiceConfigWReq) MarshalBinary() (res []byte, err error) {
+	log.Debugln("In MarshalBinary for RQueryServiceConfigWReq")
+
+	var ret []byte
+	w := bytes.NewBuffer(ret)
+
+	if len(self.ServiceHandle) != 20 {
+		return nil, fmt.Errorf("Invalid size of ServiceHandle!")
+	}
+
+	_, err = w.Write(self.ServiceHandle[:20])
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	err = binary.Write(w, le, self.BufSize)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	return w.Bytes(), nil
+}
+
+func (self *RQueryServiceConfigWReq) UnmarshalBinary(buf []byte) error {
+	return fmt.Errorf("NOT IMPLEMENTED UnmarshalBinary of RQueryServiceConfigWReq")
+}
+
+func (self *RQueryServiceConfigWRes) MarshalBinary() ([]byte, error) {
+	return nil, fmt.Errorf("NOT IMPLEMENTED MarshalBinary of RQueryServiceConfigWRes")
+}
+
+func (self *RQueryServiceConfigWRes) UnmarshalBinary(buf []byte) (err error) {
+	log.Debugln("In UnmarshalBinary for RQueryServiceConfigWRes")
+	if len(buf) < 44 {
+		return fmt.Errorf("Buffer to small for RQueryServiceConfigWRes")
+	}
+	r := bytes.NewReader(buf)
+
+	conf := QueryServiceConfigW{}
+	self.ServiceConfig = &conf
+
+	err = binary.Read(r, le, &conf.ServiceType)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+	err = binary.Read(r, le, &conf.StartType)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+	err = binary.Read(r, le, &conf.ErrorControl)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+	if len(buf) < 45 {
+		// Probably ErrorInsuficientBuffer
+		// Expect the rest of the QueryServiceConfigW struct fields to be null ptrs, so skip them
+		_, err = r.Seek(24, io.SeekCurrent)
+		if err != nil {
+			log.Errorln(err)
+			return
+		}
+
+		err = binary.Read(r, le, &self.BytesNeeded)
+		if err != nil {
+			log.Errorln(err)
+			return
+		}
+		err = binary.Read(r, le, &self.ErrorCode)
+		if err != nil {
+			log.Errorln(err)
+			return
+		}
+		return
+	}
+
+	// Skip ReferentId ptr for BinaryPathName and LoadOrderGroup
+	_, err = r.Seek(8, io.SeekCurrent)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	err = binary.Read(r, le, &conf.TagId)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	// Skip ReferentId ptr for Dependencies, ServiceStartName and DisplayName
+	_, err = r.Seek(12, io.SeekCurrent)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	conf.BinaryPathName, err = readConformantVaryingString(r)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	conf.LoadOrderGroup, err = readConformantVaryingString(r)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	conf.Dependencies, err = readConformantVaryingString(r)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	conf.ServiceStartName, err = readConformantVaryingString(r)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	conf.DisplayName, err = readConformantVaryingString(r)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	err = binary.Read(r, le, &self.BytesNeeded)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	err = binary.Read(r, le, &self.ErrorCode)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	return nil
+}
+
+func (self *RChangeServiceConfigWReq) MarshalBinary() (res []byte, err error) {
+	log.Debugln("In MarshalBinary for RChangeServiceConfigWReq")
+
+	var ret []byte
+	w := bytes.NewBuffer(ret)
+	refId := uint32(1)
+
+	if len(self.ServiceHandle) != 20 {
+		return nil, fmt.Errorf("Invalid size of ServiceHandle!")
+	}
+
+	_, err = w.Write(self.ServiceHandle[:20])
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	err = binary.Write(w, le, self.ServiceType)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	err = binary.Write(w, le, self.StartType)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	err = binary.Write(w, le, self.ErrorControl)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	_, err = writeConformantVaryingStringPtr(w, self.BinaryPathName, refId)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+	refId++
+
+	_, err = writeConformantVaryingStringPtr(w, self.LoadOrderGroup, refId)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+	refId++
+
+	err = binary.Write(w, le, self.TagId)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	//NOTE Is this correct?
+	_, err = writeConformantVaryingStringPtr(w, self.Dependencies, refId)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+	refId++
+
+	//NOTE Is this correct?
+	err = binary.Write(w, le, self.DependSize)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	_, err = writeConformantVaryingStringPtr(w, self.ServiceStartName, refId)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+	refId++
+
+	if self.Password != nil {
+		_, err = writeConformantArrayPtr(w, self.Password, refId)
+		if err != nil {
+			log.Errorln(err)
+			return
+		}
+		refId++
+
+		err = binary.Write(w, le, uint32(len(self.Password)))
+		if err != nil {
+			log.Errorln(err)
+			return
+		}
+	} else {
+		err = binary.Write(w, le, uint64(0))
+		if err != nil {
+			log.Errorln(err)
+			return
+		}
+	}
+
+	_, err = writeConformantVaryingStringPtr(w, self.DisplayName, refId)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	return w.Bytes(), nil
+}
+
+func (self *RChangeServiceConfigWReq) UnmarshalBinary(buf []byte) error {
+	return fmt.Errorf("NOT IMPLEMENTED UnmarshalBinary of RChangeServiceConfigWReq")
+}
+
+func (self *RChangeServiceConfigWRes) MarshalBinary() ([]byte, error) {
+	return nil, fmt.Errorf("NOT IMPLEMENTED MarshalBinary of RChangeServiceConfigWRes")
+}
+
+func (self *RChangeServiceConfigWRes) UnmarshalBinary(buf []byte) (err error) {
+	log.Debugln("In UnmarshalBinary for RChangeServiceConfigWRes")
+	if len(buf) < 8 {
+		return fmt.Errorf("Buffer to small for RchangeServiceConfigWRes")
+	}
+	r := bytes.NewReader(buf)
+
+	err = binary.Read(r, le, &self.TagId)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+	err = binary.Read(r, le, &self.ReturnCode)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	return nil
+}
+
+func (s *RStartServiceWRequest) MarshalBinary(meta *encoder.Metadata) ([]byte, error) {
+	var ret []byte
+	w := bytes.NewBuffer(ret)
+
+	// Encode ServiceHandle
+	w.Write(s.ServiceHandle)
+	// Encode Argc
+	if err := binary.Write(w, binary.LittleEndian, s.Argc); err != nil {
+		return nil, err
+	}
+
+	// If Argc is 0, Argv will be a null pointer
+	if s.Argc == 0 {
+		w.Write([]byte{0x0, 0x0, 0x0, 0x0})
+		return w.Bytes(), nil
+	}
+
+	refId := make([]byte, 4)
+	_, err := rand.Read(refId)
+	if err != nil {
+		log.Errorln(err)
+		return nil, err
+	}
+
+	_, err = w.Write(refId)
+	if err != nil {
+		log.Errorln(err)
+		return nil, err
+	}
+
+	// Encode Max Element Count in array
+	if err := binary.Write(w, binary.LittleEndian, s.Argc); err != nil {
+		return nil, err
+	}
+
+	// Encode another RefId for each element in the array
+	// This is because according to NDR, the pointers are lifted outside the array element and into the parent structure
+	for i := 0; i < int(s.Argc); i++ {
+		_, err := rand.Read(refId)
+		if err != nil {
+			log.Errorln(err)
+			return nil, err
+		}
+
+		_, err = w.Write(refId)
+		if err != nil {
+			log.Errorln(err)
+			return nil, err
+		}
+	}
+
+	for i := 0; i < int(s.Argc); i++ {
+		// Encode MaxCount
+		if err := binary.Write(w, binary.LittleEndian, s.Argv[i].MaxCount); err != nil {
+			log.Errorln(err)
+			return nil, err
+		}
+		// Encode Offset
+		if err := binary.Write(w, binary.LittleEndian, s.Argv[i].Offset); err != nil {
+			log.Errorln(err)
+			return nil, err
+		}
+		// Encode ActualCount
+		if err := binary.Write(w, binary.LittleEndian, s.Argv[i].ActualCount); err != nil {
+			log.Errorln(err)
+			return nil, err
+		}
+		// Encode String
+		if err := binary.Write(w, binary.LittleEndian, s.Argv[i].EncodedString); err != nil {
+			log.Errorln(err)
+			return nil, err
+		}
+		// Encode Padding
+		if len(s.Argv[i].Padd) > 0 {
+			if err := binary.Write(w, binary.LittleEndian, s.Argv[i].Padd); err != nil {
+				log.Errorln(err)
+				return nil, err
+			}
+		}
+	}
+
+	return w.Bytes(), nil
+}
+
+func (self *RStartServiceWRequest) UnmarshalBinary(buf []byte, meta *encoder.Metadata) error {
+	return fmt.Errorf("NOT IMPLEMENTED UnmarshalBinary of RStartServiceWRequest")
+}
+
+func decodeServiceConfig(config *QueryServiceConfigW) (res ServiceConfig, err error) {
+	log.Debugln("In decodeServiceConfig")
+	if _, ok := ServiceTypeStatusMap[config.ServiceType]; !ok {
+		err = fmt.Errorf("Could not identify returned service type: %d\n", config.ServiceType)
+		fmt.Println(err)
+	}
+	res.ServiceType = ServiceTypeStatusMap[config.ServiceType]
+
+	if _, ok := StartTypeStatusMap[config.StartType]; !ok {
+		err = fmt.Errorf("Could not identify returned start type: %d\n", config.StartType)
+		fmt.Println(err)
+	}
+	res.StartType = StartTypeStatusMap[config.StartType]
+
+	if _, ok := ErrorControlStatusMap[config.ErrorControl]; !ok {
+		err = fmt.Errorf("Could not identify returned start type: %d\n", config.ErrorControl)
+		fmt.Println(err)
+	}
+	res.ErrorControl = ErrorControlStatusMap[config.ErrorControl]
+
+	res.BinaryPathName = config.BinaryPathName
+	if err != nil {
+		err = fmt.Errorf("Error decoding service config: %s\n", err)
+		fmt.Println(err)
+	}
+
+	res.LoadOrderGroup = config.LoadOrderGroup
+	if err != nil {
+		err = fmt.Errorf("Error decoding service config: %s\n", err)
+		fmt.Println(err)
+	}
+
+	res.TagId = config.TagId
+
+	res.Dependencies = config.Dependencies
+	if err != nil {
+		err = fmt.Errorf("Error decoding service config: %s\n", err)
+		fmt.Println(err)
+	}
+
+	res.ServiceStartName = config.ServiceStartName
+	if err != nil {
+		err = fmt.Errorf("Error decoding service config: %s\n", err)
+		fmt.Println(err)
+	}
+
+	res.DisplayName = config.DisplayName
+	if err != nil {
+		err = fmt.Errorf("Error decoding service config: %s\n", err)
+		fmt.Println(err)
+	}
+
+	return
+}
+
+func (sb *ServiceBind) openSCManager(desiredAccess uint32) (handle []byte, err error) {
+	log.Debugln("In openSCManager")
+	scReq := ROpenSCManagerWRequest{
+		MachineName:   "DUMMY",
+		DatabaseName:  "ServicesActive",
+		DesiredAccess: desiredAccess,
+	}
+	scBuf, err := scReq.MarshalBinary()
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	buffer, err := sb.MakeIoCtlRequest(SvcCtlROpenSCManagerW, scBuf)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	// Retrieve context handle from response
+	res := ROpenSCManagerWResponse{}
+	err = res.UnmarshalBinary(buffer)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	if res.ReturnCode != ErrorSuccess {
+		err = ServiceResponseCodeMap[res.ReturnCode]
+	}
+
+	handle = res.ContextHandle[:]
+	return
+}
+
+func (sb *ServiceBind) openService(scHandle []byte, serviceName string, desiredAccess uint32) (handle []byte, err error) {
+	log.Debugln("In openService")
+	serviceReq := ROpenServiceWRequest{
+		SCContextHandle: scHandle,
+		ServiceName:     serviceName,
+		DesiredAccess:   desiredAccess,
+	}
+
+	serviceBuf, err := serviceReq.MarshalBinary()
+	if err != nil {
+		return
+	}
+
+	buffer, err := sb.MakeIoCtlRequest(SvcCtlROpenServiceW, serviceBuf)
+	if err != nil {
+		return
+	}
+
+	// Retrieve context handle from response
+	res := ROpenServiceWResponse{}
+	err = res.UnmarshalBinary(buffer)
+	if err != nil {
+		return
+	}
+
+	if res.ReturnCode != ErrorSuccess {
+		err = ServiceResponseCodeMap[res.ReturnCode]
+	}
+
+	handle = res.ContextHandle
+	return
+}
+
+func (sb *ServiceBind) GetServiceStatus(serviceName string) (status uint32, err error) {
+	log.Debugln("In GetServiceStatus")
+	handle, err := sb.openSCManager(ServiceQueryStatus)
+	if err != nil {
+		return
+	}
+	defer sb.CloseServiceHandle(handle)
+
+	serviceHandle, err := sb.openService(handle, serviceName, ServiceQueryStatus)
+	if err != nil {
+		return
+	}
+	defer sb.CloseServiceHandle(serviceHandle)
+
+	ssReq := RQueryServiceStatusReq{ContextHandle: serviceHandle}
+	ssBuf, err := ssReq.MarshalBinary()
+	if err != nil {
+		return
+	}
+
+	buffer, err := sb.MakeIoCtlRequest(SvcCtlRQueryServiceStatus, ssBuf)
+	if err != nil {
+		return
+	}
+
+	// Retrieve context handle from response
+	res := RQueryServiceStatusRes{}
+	err = res.UnmarshalBinary(buffer)
+	if err != nil {
+		return
+	}
+
+	if res.ReturnCode != ErrorSuccess {
+		err = ServiceResponseCodeMap[res.ReturnCode]
+		return
+	}
+
+	status = res.ServiceStatus.CurrentState
+	return
+}
+
+func (sb *ServiceBind) StartService(serviceName string, args []string) (err error) {
+	log.Debugln("In StartService")
+	handle, err := sb.openSCManager(ServiceStart)
+	if err != nil {
+		return
+	}
+	defer sb.CloseServiceHandle(handle)
+	serviceHandle, err := sb.openService(handle, serviceName, ServiceStart)
+	if err != nil {
+		return
+	}
+	defer sb.CloseServiceHandle(serviceHandle)
+
+	ssReq := RStartServiceWRequest{ServiceHandle: serviceHandle}
+	if args == nil || len(args) == 0 {
+		ssReq.Argc = 0
+		ssReq.Argv = nil
+	} else {
+		arr := make([]UnicodeStr, len(args))
+		refId := uint32(1)
+		ssReq.Argc = uint32(len(args))
+		for i := range args {
+			arr[i] = *NewUnicodeStr(refId, args[i])
+			refId++
+		}
+		ssReq.Argv = arr
+	}
+
+	ssBuf, err := encoder.Marshal(&ssReq)
+	if err != nil {
+		return
+	}
+
+	buffer, err := sb.MakeIoCtlRequest(SvcCtlRStartServiceW, ssBuf)
+	if err != nil {
+		return
+	}
+
+	// Retrieve context handle from response
+	returnValue := binary.LittleEndian.Uint32(buffer)
+	if returnValue != ErrorSuccess {
+		return ServiceResponseCodeMap[returnValue]
+	}
+
+	return
+}
+
+func (sb *ServiceBind) ControlService(serviceName string, control uint32) (err error) {
+	log.Debugln("In ControlService")
+	handle, err := sb.openSCManager(SCManagerConnect)
+	if err != nil {
+		return
+	}
+	defer sb.CloseServiceHandle(handle)
+	serviceHandle, err := sb.openService(handle, serviceName, ServicePauseContinue|ServiceInterrogate|ServiceStop)
+	if err != nil {
+		return
+	}
+	defer sb.CloseServiceHandle(serviceHandle)
+
+	csReq := RControlServiceRequest{
+		ServiceHandle: serviceHandle,
+		Control:       control,
+	}
+	csBuf, err := encoder.Marshal(csReq)
+	if err != nil {
+		return
+	}
+
+	buffer, err := sb.MakeIoCtlRequest(SvcCtlRControlService, csBuf)
+	if err != nil {
+		return
+	}
+
+	// Parse ServiceStatus
+	res := RControlServiceResponse{}
+	err = encoder.Unmarshal(buffer, &res)
+	if err != nil {
+		return
+	}
+
+	// Retrieve context handle from response
+	if res.ReturnValue != ErrorSuccess {
+		return ServiceResponseCodeMap[res.ReturnValue]
+	}
+
+	return
+}
+
+func (sb *ServiceBind) GetServiceConfig(serviceName string) (config ServiceConfig, err error) {
+	log.Debugln("In GetServiceConfig")
+	handle, err := sb.openSCManager(ServiceQueryConfig)
+	if err != nil {
+		return
+	}
+	defer sb.CloseServiceHandle(handle)
+	serviceHandle, err := sb.openService(handle, serviceName, ServiceQueryConfig)
+	if err != nil {
+		return
+	}
+	defer sb.CloseServiceHandle(serviceHandle)
+
+	innerReq := RQueryServiceConfigWReq{
+		ServiceHandle: serviceHandle,
+		BufSize:       0,
+	}
+	innerBuf, err := innerReq.MarshalBinary()
+	if err != nil {
+		return
+	}
+
+	// Make request to figure out buffer size
+	buffer, err := sb.MakeIoCtlRequest(SvcCtlRQueryServiceConfigW, innerBuf)
+	if err != nil {
+		return
+	}
+
+	// Parse response
+	res := RQueryServiceConfigWRes{}
+	err = res.UnmarshalBinary(buffer)
+	if err != nil {
+		return
+	}
+
+	if res.ErrorCode != ErrorInsufficientBuffer {
+		err = ServiceResponseCodeMap[res.ErrorCode]
+		return
+	}
+
+	// Repeat request with allocated buffer size
+	innerReq.BufSize = res.BytesNeeded
+	innerBuf2, err := innerReq.MarshalBinary()
+	if err != nil {
+		return
+	}
+
+	buffer, err = sb.MakeIoCtlRequest(SvcCtlRQueryServiceConfigW, innerBuf2)
+	if err != nil {
+		return
+	}
+
+	// Parse ServiceConfig
+	res = RQueryServiceConfigWRes{}
+	err = res.UnmarshalBinary(buffer)
+	if err != nil {
+		return
+	}
+
+	if res.ErrorCode != ErrorSuccess {
+		err = ServiceResponseCodeMap[res.ErrorCode]
+		return
+	}
+
+	return decodeServiceConfig(res.ServiceConfig)
+}
+
+func (sb *ServiceBind) ChangeServiceConfig(
+	serviceName string,
+	serviceType, startType, errorControl uint32,
+	binaryPathName, serviceStartName, password, displayName string) (err error) {
+
+	log.Debugln("In ChangeServiceConfig")
+
+	handle, err := sb.openSCManager(SCManagerEnumerateService)
+	if err != nil {
+		return
+	}
+	defer sb.CloseServiceHandle(handle)
+	serviceHandle, err := sb.openService(handle, serviceName, ServiceChangeConfig)
+	if err != nil {
+		return
+	}
+	defer sb.CloseServiceHandle(serviceHandle)
+
+	innerReq := RChangeServiceConfigWReq{
+		ServiceHandle:    serviceHandle,
+		ServiceType:      serviceType,
+		StartType:        startType,
+		ErrorControl:     errorControl,
+		BinaryPathName:   binaryPathName,
+		LoadOrderGroup:   "",
+		TagId:            0,
+		Dependencies:     "",
+		DependSize:       0,
+		ServiceStartName: serviceStartName,
+		DisplayName:      displayName,
+	}
+
+	/*
+	   MS-SCMR Section 3.1.4.12 explains that in RPC over TCP, the password should be plaintext,
+	   but over SMB it must be encrypted
+	*/
+	if password != "" {
+		uncPassword := ToUnicode(password + "\x00")
+		encPassword, err := encryptSecret(sb.f.GetSessionKey(), uncPassword)
+		if err != nil {
+			log.Errorln(err)
+			return err
+		}
+		innerReq.Password = encPassword
+	}
+
+	innerBuf, err := innerReq.MarshalBinary()
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	buffer, err := sb.MakeIoCtlRequest(SvcCtlRChangeServiceConfigW, innerBuf)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	// Parse ServiceConfig
+	res := RChangeServiceConfigWRes{}
+	err = res.UnmarshalBinary(buffer)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+	if res.ReturnCode != ErrorSuccess {
+		status, found := ServiceResponseCodeMap[res.ReturnCode]
+		if !found {
+			err = fmt.Errorf("Received unknown return code for RChangeServiceConfigW: 0x%x\n", res.ReturnCode)
+			log.Errorln(err)
+			return err
+		}
+		return status
+	}
+
+	if res.ReturnCode != ErrorSuccess {
+
+		return ServiceResponseCodeMap[res.ReturnCode]
+	}
+
+	return
+}
+
+func (sb *ServiceBind) CreateService(
+	serviceName string,
+	serviceType, startType, errorControl uint32,
+	binaryPathName, serviceStartName, displayName string, startService bool) (err error) {
+
+	log.Debugln("In CreateService")
+
+	scHandle, err := sb.openSCManager(SCManagerCreateService)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+	defer sb.CloseServiceHandle(scHandle)
+
+	innerReq := RCreateServiceRequest{
+		SCContextHandle:  scHandle,
+		ServiceName:      NewUnicodeStr(0, serviceName),
+		DisplayName:      NewUnicodeStr(1, displayName),
+		DesiredAccess:    ServiceAllAccess,
+		ServiceType:      serviceType,
+		StartType:        startType,
+		ErrorControl:     errorControl,
+		BinaryPathName:   NewUnicodeStr(0, binaryPathName),
+		LoadOrderGroup:   nil,
+		TagId:            0,
+		Dependencies:     nil,
+		DependSize:       0,
+		ServiceStartName: NewUnicodeStr(2, serviceStartName),
+	}
+
+	log.Debugf("ServiceName: %s\n", innerReq.ServiceName.EncodedString)
+	// To support specifying a password I must figure out how the encryption is
+	// performed as this parameter expects an encrypted passphrase with some
+	// session key
+	//if password != "" {
+	//    innerReq.Password = NewUnicodeStr(0, password)
+	//    innerReq.PwSize = uint32((len(password)+1) * 2) // Null-terminated unicode string
+	//}
+
+	innerBuf, err := encoder.Marshal(innerReq)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	buffer, err := sb.MakeIoCtlRequest(SvcCtlRCreateServiceW, innerBuf)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+
+	// Parse ServiceConfig
+	res := RCreateServiceResponse{}
+	err = encoder.Unmarshal(buffer, &res)
+	if err != nil {
+		log.Errorln(err)
+		return
+	}
+	if res.ReturnCode != ErrorSuccess {
+		err = ServiceResponseCodeMap[res.ReturnCode]
+		return
+	}
+	defer sb.CloseServiceHandle(res.ContextHandle)
+
+	if startService {
+		ssReq := RStartServiceWRequest{ServiceHandle: res.ContextHandle}
+		ssReq.Argc = 0
+		// When Argc is 0 I need to marshal 0x00000000 for Argc and same for Argv e.g., 4 bytes combined of 0s
+
+		ssBuf, err2 := encoder.Marshal(&ssReq)
+		if err != nil {
+			return err2
+		}
+
+		buffer, err2 := sb.MakeIoCtlRequest(SvcCtlRStartServiceW, ssBuf)
+		if err != nil {
+			return err2
+		}
+
+		returnValue := binary.LittleEndian.Uint32(buffer)
+		if returnValue != ErrorSuccess {
+			return ServiceResponseCodeMap[returnValue]
+		}
+	}
+
+	return
+}
+
+func (sb *ServiceBind) DeleteService(serviceName string) (err error) {
+	scHandle, err := sb.openSCManager(SCManagerConnect)
+	if err != nil {
+		return
+	}
+	defer sb.CloseServiceHandle(scHandle)
+	handle, err := sb.openService(scHandle, serviceName, ServiceAllAccess)
+	if err != nil {
+		return
+	}
+	defer sb.CloseServiceHandle(handle)
+
+	// Attempt to stop the service before deletion
+	csReq := RControlServiceRequest{
+		ServiceHandle: handle,
+		Control:       ServiceControlStop,
+	}
+	csBuf, err := encoder.Marshal(csReq)
+	if err != nil {
+		return
+	}
+
+	_, err = sb.MakeIoCtlRequest(SvcCtlRControlService, csBuf)
+	if err != nil {
+		log.Errorln(err)
+		// Continue with deletion even if stop failed for some reason
+	}
+
+	innerReq := RDeleteServiceRequest{
+		ServiceHandle: handle,
+	}
+
+	innerBuf, err := encoder.Marshal(innerReq)
+	if err != nil {
+		return
+	}
+
+	buffer, err := sb.MakeIoCtlRequest(SvcCtlRDeleteService, innerBuf)
+	if err != nil {
+		return
+	}
+
+	// Parse ServiceConfig
+	res := RDeleteServiceResponse{}
+	err = encoder.Unmarshal(buffer, &res)
+	if err != nil {
+		return
+	}
+	if res.ReturnCode != ErrorSuccess {
+		err = ServiceResponseCodeMap[res.ReturnCode]
+		return
+	}
+	return
+}
+
+func (sb *ServiceBind) CloseServiceHandle(serviceHandle []byte) {
+	//log.Debugln("In CloseServiceHandle")
+	closeReq := RCloseServiceHandleReq{
+		ServiceHandle: serviceHandle,
+	}
+	closeBuf, err := closeReq.MarshalBinary()
+	if err != nil {
+		fmt.Printf("Failed to encode close service handle request")
+		return
+	}
+
+	buffer, err := sb.MakeIoCtlRequest(SvcCtlRCloseServiceHandle, closeBuf)
+	if err != nil {
+		fmt.Printf("Failed to close service handle with error: %s\n", err)
+		return
+	}
+	res := RCloseServiceHandleRes{}
+	err = res.UnmarshalBinary(buffer)
+	if err != nil {
+		fmt.Printf("Failed to unmarshal response of close service handle")
+		return
+	}
+
+	if res.ReturnCode != ErrorSuccess {
+		log.Errorf("Failed to close service handle with error (return value: 0x%x): %v\n", res.ReturnCode, ServiceResponseCodeMap[res.ReturnCode])
+	}
+
+	return
+}
