+++ smb/dcerpc/scmr_test.go
@@ -30,245 +30,6 @@ import (
 	"testing"
 )
 
-func TestBindReq(t *testing.T) {
-	// Simple test to verify that the packet structure is valid
-	pkt, err := hex.DecodeString("05000b0310000000480000004204cb9ab810b81000000000010000000000010081bb7a364498f135ad3298f03800100302000000045d888aeb1cc9119fe808002b10486002000000")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	req, err := NewBindReq(2596996162, "367abb81-9844-35f1-ad32-98f038001003", 2, 0, "8a885d04-1ceb-11c9-9fe8-08002b104860")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	buf, err := encoder.Marshal(req)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if !bytes.Equal(pkt, buf) {
-		t.Error("Fail")
-	}
-}
-
-func TestBindRes(t *testing.T) {
-	// Simple test to verify that the packet structure is valid
-	resPkt, err := hex.DecodeString("05000c0310000000440000004204cb9ab810b810d75400000d005c706970655c6e747376637300000100000000000000045d888aeb1cc9119fe808002b10486002000000")
-	if err != nil {
-		t.Fatal(err)
-	}
-	var res BindRes = NewBindRes()
-	err = encoder.Unmarshal(resPkt, &res)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if res.MajorVersion != 5 {
-		t.Error("Fail")
-	}
-
-	if res.Flags != 3 {
-		t.Error("Fail")
-	}
-
-	if res.Representation != 16 {
-		t.Error("Fail")
-	}
-
-	if res.FragLength != 68 {
-		t.Error("Fail")
-	}
-
-	if res.CallId != 2596996162 {
-		t.Error("Fail")
-	}
-
-	if res.MaxRecvFragSize != 4280 {
-		t.Error("Fail")
-	}
-
-	if res.MaxSendFragSize != 4280 {
-		t.Error("Fail")
-	}
-
-	if res.Association != 0x000054d7 {
-		t.Error("Fail")
-	}
-
-	if res.SecAddrLen != 13 {
-		t.Error("Fail")
-	}
-
-	if !bytes.Equal(res.SecAddr, []byte("\\pipe\\ntsvcs\x00")) {
-		t.Error("Fail")
-	}
-
-	if res.CtxCount != 1 {
-		t.Error("Fail")
-	}
-	var items ContextResItems = *res.Context
-
-	if items[0].Result != 0 {
-		t.Error("Fail")
-	}
-
-	ndr, err := hex.DecodeString("045d888aeb1cc9119fe808002b104860")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if !bytes.Equal(items[0].TransferUUID, ndr) {
-		t.Error("Fail")
-	}
-
-	if items[0].TransferVersion != 2 {
-		t.Error("Fail")
-	}
-}
-
-func TestNetShareEnumAllReq(t *testing.T) {
-	// Simple test to verify that the packet structure is valid
-	pkt, err := hex.DecodeString("000002001100000000000000110000004400450053004b0054004f0050002d0041004900470030004300310044003200000000000100000001000000020000000000000000000000ffffffff0800020000000000")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	req := NewNetShareEnumAllRequest("DESKTOP-AIG0C1D2")
-
-	buf, err := encoder.Marshal(req)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if !bytes.Equal(pkt, buf) {
-		t.Error("Fail")
-	}
-}
-
-func TestNetShareEnumAllRes(t *testing.T) {
-	// Simple test to verify that the packet structure is valid
-	resPkt, err := hex.DecodeString("01000000010000000000020005000000040002000500000008000200000000800c00020010000200000000801400020018000200030000801c00020020000200000000002400020028000200000000002c000200070000000000000007000000410044004d0049004e002400000000000d000000000000000d000000520065006d006f00740065002000410064006d0069006e000000000003000000000000000300000043002400000000000e000000000000000e000000440065006600610075006c00740020007300680061007200650000000500000000000000050000004900500043002400000000000b000000000000000b000000520065006d006f00740065002000490050004300000000000900000000000000090000006d00790066006f006c0064006500720000000000010000000000000001000000000000000600000000000000060000005500730065007200730000000100000000000000010000000000000005000000300002000000000000000000")
-	if err != nil {
-		t.Fatal(err)
-	}
-	res := NetShareEnumAllResponse{}
-	err = encoder.Unmarshal(resPkt, &res)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if res.Level != 1 {
-		t.Error("Fail")
-	}
-
-	if res.NetShareCtr.Ctr != 1 {
-		t.Error("Fail")
-	}
-
-	ctr1 := res.NetShareCtr.Pointer.(*NetShareCtr1)
-
-	if ctr1.Count != 5 {
-		t.Error("Fail")
-	}
-
-	if ctr1.Info.MaxCount != 5 {
-		t.Error("Fail")
-	}
-
-	if ctr1.Pointer[0].Type != StypeDisktree|StypeSpecial {
-		t.Error("Fail")
-	}
-
-	if val, err := encoder.FromUnicodeString(ctr1.Pointer[0].Name.EncodedString); err != nil {
-		t.Error("Fail")
-	} else if val != "ADMIN$\x00" {
-		t.Error("Fail")
-	}
-
-	if val, err := encoder.FromUnicodeString(ctr1.Pointer[0].Comment.EncodedString); err != nil {
-		t.Error("Fail")
-	} else if val != "Remote Admin\x00" {
-		t.Error("Fail")
-	}
-
-	if ctr1.Pointer[1].Type != StypeDisktree|StypeSpecial {
-		t.Error("Fail")
-	}
-
-	if val, err := encoder.FromUnicodeString(ctr1.Pointer[1].Name.EncodedString); err != nil {
-		t.Error("Fail")
-	} else if val != "C$\x00" {
-		t.Error("Fail")
-	}
-
-	if val, err := encoder.FromUnicodeString(ctr1.Pointer[1].Comment.EncodedString); err != nil {
-		t.Error("Fail")
-	} else if val != "Default share\x00" {
-		t.Error("Fail")
-	}
-
-	if ctr1.Pointer[2].Type != StypeIPC|StypeSpecial {
-		t.Error("Fail")
-	}
-
-	if val, err := encoder.FromUnicodeString(ctr1.Pointer[2].Name.EncodedString); err != nil {
-		t.Error("Fail")
-	} else if val != "IPC$\x00" {
-		t.Error("Fail")
-	}
-
-	if val, err := encoder.FromUnicodeString(ctr1.Pointer[2].Comment.EncodedString); err != nil {
-		t.Error("Fail")
-	} else if val != "Remote IPC\x00" {
-		t.Error("Fail")
-	}
-
-	if ctr1.Pointer[3].Type != StypeDisktree {
-		t.Error("Fail")
-	}
-
-	if val, err := encoder.FromUnicodeString(ctr1.Pointer[3].Name.EncodedString); err != nil {
-		t.Error("Fail")
-	} else if val != "myfolder\x00" {
-		t.Error("Fail")
-	}
-
-	if val, err := encoder.FromUnicodeString(ctr1.Pointer[3].Comment.EncodedString); err != nil {
-		t.Error("Fail")
-	} else if val != "\x00" {
-		t.Error("Fail")
-	}
-
-	if ctr1.Pointer[4].Type != StypeDisktree {
-		t.Error("Fail")
-	}
-
-	if val, err := encoder.FromUnicodeString(ctr1.Pointer[4].Name.EncodedString); err != nil {
-		t.Error("Fail")
-	} else if val != "Users\x00" {
-		t.Error("Fail")
-	}
-
-	if val, err := encoder.FromUnicodeString(ctr1.Pointer[4].Comment.EncodedString); err != nil {
-		t.Error("Fail")
-	} else if val != "\x00" {
-		t.Error("Fail")
-	}
-
-	if res.TotalEntries != 5 {
-		t.Error("Fail")
-	}
-
-	if res.ResumeHandle.Handle != 0 {
-		t.Error("Fail")
-	}
-
-	if res.WindowsError != 0 {
-		t.Error("Fail")
-	}
-}
-
 func TestOpenSCManagerReq(t *testing.T) {
 	// Simple test to verify that the packet structure is valid
 	pkt, err := hex.DecodeString("01000000060000000000000006000000440055004d004d0059000000020000000f000000000000000f000000530065007200760069006300650073004100630074006900760065000000000004000000")
@@ -277,17 +38,17 @@ func TestOpenSCManagerReq(t *testing.T) {
 	}
 
 	req := ROpenSCManagerWRequest{
-		MachineName:   *NewUnicodeStr(1, "DUMMY"),
-		DatabaseName:  *NewUnicodeStr(2, "ServicesActive"),
+		MachineName:   "DUMMY",
+		DatabaseName:  "ServicesActive",
 		DesiredAccess: 4,
 	}
-	buf, err := encoder.Marshal(req)
+	buf, err := req.MarshalBinary()
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	if !bytes.Equal(pkt, buf) {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 }
 
@@ -303,17 +64,17 @@ func TestOpenSCManagerRes(t *testing.T) {
 	}
 
 	res := ROpenSCManagerWResponse{}
-	err = encoder.Unmarshal(resPkt, &res)
+	err = res.UnmarshalBinary(resPkt)
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	if !bytes.Equal(res.ContextHandle, handle) {
-		t.Error("Fail")
+	if !bytes.Equal(res.ContextHandle[:], handle) {
+		t.Fatal("Fail")
 	}
 
 	if res.ReturnCode != 0 {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 }
 
@@ -331,16 +92,16 @@ func TestOpenServiceReq(t *testing.T) {
 
 	req := ROpenServiceWRequest{
 		SCContextHandle: handle,
-		ServiceName:     NewUnicodeStr(0, "RemoteRegistry"),
+		ServiceName:     "RemoteRegistry",
 		DesiredAccess:   4,
 	}
-	buf, err := encoder.Marshal(req)
+	buf, err := req.MarshalBinary()
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	if !bytes.Equal(pkt, buf) {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 }
 
@@ -356,17 +117,17 @@ func TestOpenServiceRes(t *testing.T) {
 	}
 
 	res := ROpenServiceWResponse{}
-	err = encoder.Unmarshal(resPkt, &res)
+	err = res.UnmarshalBinary(resPkt)
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	if !bytes.Equal(res.ContextHandle, handle) {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 
 	if res.ReturnCode != 0 {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 }
 
@@ -382,14 +143,14 @@ func TestQueryServiceStatusReq(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	req := RQueryServiceStatusRequest{ContextHandle: handle}
-	buf, err := encoder.Marshal(req)
+	req := RQueryServiceStatusReq{ContextHandle: handle}
+	buf, err := req.MarshalBinary()
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	if !bytes.Equal(pkt, buf) {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 }
 
@@ -400,42 +161,42 @@ func TestQueryServiceStatusRes(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	res := RQueryServiceStatusResponse{}
-	err = encoder.Unmarshal(resPkt, &res)
+	res := RQueryServiceStatusRes{}
+	err = res.UnmarshalBinary(resPkt)
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	if res.ServiceType != ServiceWin32OwnProcess|ServiceWin32ShareProcess {
-		t.Error("Fail")
+	if res.ServiceStatus.ServiceType != ServiceWin32OwnProcess|ServiceWin32ShareProcess {
+		t.Fatal("Fail")
 	}
 
-	if res.CurrentState != ServiceRunning {
-		t.Error("Fail")
+	if res.ServiceStatus.CurrentState != ServiceRunning {
+		t.Fatal("Fail")
 	}
 
-	if res.ControlsAccepted != ServiceControlStop {
-		t.Error("Fail")
+	if res.ServiceStatus.ControlsAccepted != ServiceControlStop {
+		t.Fatal("Fail")
 	}
 
-	if res.Win32ExitCode != 0 {
-		t.Error("Fail")
+	if res.ServiceStatus.Win32ExitCode != 0 {
+		t.Fatal("Fail")
 	}
 
-	if res.ServiceSpecificExitCode != 0 {
-		t.Error("Fail")
+	if res.ServiceStatus.ServiceSpecificExitCode != 0 {
+		t.Fatal("Fail")
 	}
 
-	if res.CheckPoint != 0 {
-		t.Error("Fail")
+	if res.ServiceStatus.CheckPoint != 0 {
+		t.Fatal("Fail")
 	}
 
-	if res.WaitHint != 0 {
-		t.Error("Fail")
+	if res.ServiceStatus.WaitHint != 0 {
+		t.Fatal("Fail")
 	}
 
 	if res.ReturnCode != 0 {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 }
 
@@ -464,7 +225,7 @@ func TestStartServiceReq(t *testing.T) {
 	}
 
 	if !bytes.Equal(pkt, buf) {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 }
 
@@ -490,7 +251,7 @@ func TestControlServiceReq(t *testing.T) {
 	}
 
 	if !bytes.Equal(pkt, buf) {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 }
 
@@ -508,35 +269,35 @@ func TestControlServiceRes(t *testing.T) {
 	}
 
 	if res.ServiceType != ServiceWin32OwnProcess|ServiceWin32ShareProcess {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 
 	if res.CurrentState != ServiceStopPending {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 
 	if res.ControlsAccepted != ServiceControlStop {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 
 	if res.Win32ExitCode != 0x042a {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 
 	if res.ServiceSpecificExitCode != 0 {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 
 	if res.CheckPoint != 3 {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 
 	if res.WaitHint != 0x0bb8 {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 
 	if res.ReturnValue != 0 {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 }
 
@@ -552,17 +313,17 @@ func TestQueryServiceConfig(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	req := RQueryServiceConfigWRequest{
+	req := RQueryServiceConfigWReq{
 		ServiceHandle: handle,
 		BufSize:       258,
 	}
-	buf, err := encoder.Marshal(req)
+	buf, err := req.MarshalBinary()
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	if !bytes.Equal(pkt, buf) {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 }
 
@@ -573,100 +334,102 @@ func TestQueryServiceConfigRes(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	res := RQueryServiceConfigWResponse{}
-	err = encoder.Unmarshal(resPkt, &res)
+	res := RQueryServiceConfigWRes{}
+	err = res.UnmarshalBinary(resPkt)
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	if res.ErrorCode != 0 {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 
 	if res.BytesNeeded != 0x0102 {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 
 	if res.ServiceConfig.ServiceType != ServiceWin32ShareProcess {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 
 	if res.ServiceConfig.StartType != ServiceDemandStart {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 
 	if res.ServiceConfig.ErrorControl != ServiceErrorIgnore {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 
-	if val, err := encoder.FromUnicodeString(res.ServiceConfig.BinaryPathName.EncodedString); err != nil {
-		t.Error("Fail")
-	} else if val != "C:\\Windows\\System32\\svchost.exe -k rdxgroup\x00" {
-		t.Error("Fail")
+	if res.ServiceConfig.BinaryPathName != "C:\\Windows\\System32\\svchost.exe -k rdxgroup" {
+		t.Fatal("Fail")
 	}
 
-	if val, err := encoder.FromUnicodeString(res.ServiceConfig.LoadOrderGroup.EncodedString); err != nil {
-		t.Error("Fail")
-	} else if val != "\x00" {
-		t.Error("Fail")
+	if res.ServiceConfig.LoadOrderGroup != "" {
+		t.Fatal("Fail")
 	}
 
 	if res.ServiceConfig.TagId != 0 {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 
-	if val, err := encoder.FromUnicodeString(res.ServiceConfig.Dependencies.EncodedString); err != nil {
-		t.Error("Fail")
-	} else if val != "/\x00" {
-		t.Error("Fail")
+	if res.ServiceConfig.Dependencies != "/" {
+		t.Fatal("Fail")
 	}
 
-	if val, err := encoder.FromUnicodeString(res.ServiceConfig.ServiceStartName.EncodedString); err != nil {
-		t.Error("Fail")
-	} else if val != "LocalSystem\x00" {
-		t.Error("Fail")
+	if res.ServiceConfig.ServiceStartName != "LocalSystem" {
+		t.Fatal("Fail")
 	}
 
-	if val, err := encoder.FromUnicodeString(res.ServiceConfig.DisplayName.EncodedString); err != nil {
-		t.Error("Fail")
-	} else if val != "Retail Demo Service\x00" {
-		t.Error("Fail")
+	if res.ServiceConfig.DisplayName != "Retail Demo Service" {
+		t.Fatal("Fail")
 	}
 }
 
 func TestChangeServiceConfigReq(t *testing.T) {
 	// Simple test to verify that the packet structure is valid
-	pkt, err := hex.DecodeString("00000000dafba1738895c44798b2f490119e5084ffffffff04000000ffffffff010000002c000000000000002c00000043003a005c00570069006e0064006f00770073005c00530079007300740065006d00330032005c0073007600630068006f00730074002e0065007800650020002d006b002000720064007800670072006f0075007000000000000000000000000000000000000000020000000c000000000000000c0000004c006f00630061006c00530079007300740065006d000000000000000000000003000000150000000000000015000000520065007400610069006c002000440065006d006f0020005300650072007600690063006500320000000000")
+	pkt, err := hex.DecodeString("000000009b8a1554f3da95418cbe3c9f6be68d0bffffffff0400000000000000010000002c000000000000002c00000043003a005c00570069006e0064006f00770073005c00530079007300740065006d00330032005c0073007600630068006f00730074002e0065007800650020002d006b002000720064007800670072006f0075007000000000000000000000000000000000000000040000001000000000000000100000002e005c00610064006d0069006e006900730074007200610074006f007200000005000000200000003e3af69cb90cdb3b983dc9c3d7042d72e1981e344d226c6789a3237e184262e52000000006000000150000000000000015000000520065007400610069006c002000640065006d006f0020007300650072007600690063006500320000000000")
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	handle, err := hex.DecodeString("00000000dafba1738895c44798b2f490119e5084")
+	handle, err := hex.DecodeString("000000009b8a1554f3da95418cbe3c9f6be68d0b")
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	req := RChangeServiceConfigWRequest{
+	sessionKey, err := hex.DecodeString("47780a2c2a3bb5890a0b70328790829d")
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	req := RChangeServiceConfigWReq{
 		ServiceHandle:    handle,
 		ServiceType:      ServiceNoChange,
 		StartType:        ServiceDisabled,
-		ErrorControl:     ServiceNoChange,
-		BinaryPathName:   NewUnicodeStr(1, "C:\\Windows\\System32\\svchost.exe -k rdxgroup"),
-		LoadOrderGroup:   nil,
-		Dependencies:     nil,
+		ErrorControl:     ServiceErrorIgnore,
+		BinaryPathName:   "C:\\Windows\\System32\\svchost.exe -k rdxgroup",
+		LoadOrderGroup:   "",
+		Dependencies:     "",
 		DependSize:       0,
-		ServiceStartName: NewUnicodeStr(2, "LocalSystem"),
-		Password:         nil,
-		PwSize:           0,
-		DisplayName:      NewUnicodeStr(3, "Retail Demo Service2"),
+		ServiceStartName: ".\\administrator",
+		DisplayName:      "Retail demo service2",
 	}
-	buf, err := encoder.Marshal(req)
+
+	password := "secretpass"
+	uncPassword := ToUnicode(password + "\x00")
+	encPassword, err := encryptSecret(sessionKey, uncPassword)
+	if err != nil {
+		t.Fatal(err)
+	}
+	req.Password = encPassword
+
+	buf, err := req.MarshalBinary()
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	if !bytes.Equal(pkt, buf) {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 }
 
@@ -677,18 +440,18 @@ func TestChangeServiceConfigRes(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	res := RChangeServiceConfigWResponse{}
-	err = encoder.Unmarshal(resPkt, &res)
+	res := RChangeServiceConfigWRes{}
+	err = res.UnmarshalBinary(resPkt)
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	if res.TagId != 0 {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 
 	if res.ReturnCode != 0 {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 }
 
@@ -706,13 +469,13 @@ func TestCloseServiceHandleReq(t *testing.T) {
 
 	req := RCloseServiceHandleReq{ServiceHandle: handle}
 
-	buf, err := encoder.Marshal(req)
+	buf, err := req.MarshalBinary()
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	if !bytes.Equal(pkt, buf) {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 }
 
@@ -723,23 +486,23 @@ func TestCloseServiceHandleRes(t *testing.T) {
 		t.Fatal(err)
 	}
 	res := RCloseServiceHandleRes{}
-	err = encoder.Unmarshal(resPkt, &res)
+	err = res.UnmarshalBinary(resPkt)
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	if !bytes.Equal(res.ContextHandle, make([]byte, 20)) {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 
 	if res.ReturnCode != 0 {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 }
 
 func TestRCreateServiceReq(t *testing.T) {
 	// Simple test to verify that the packet structure is valid
-    pkt, err := hex.DecodeString("000000008380de25416dfe4ab011e6b3fbf620980a000000000000000a0000004d00690073006300530056004300310032000000010000000a000000000000000a0000004d00690073006300530056004300310032000000ff010f001000000003000000000000001d000000000000001d00000043003a005c00770069006e0064006f00770073005c00740065006d0070005c006f006e006500640072006900760065002e006500780065000000000000000000000000000000000000000000020000000c000000000000000c0000004c006f00630061006c00530079007300740065006d0000000000000000000000")
+	pkt, err := hex.DecodeString("000000008380de25416dfe4ab011e6b3fbf620980a000000000000000a0000004d00690073006300530056004300310032000000010000000a000000000000000a0000004d00690073006300530056004300310032000000ff010f001000000003000000000000001d000000000000001d00000043003a005c00770069006e0064006f00770073005c00740065006d0070005c006f006e006500640072006900760065002e006500780065000000000000000000000000000000000000000000020000000c000000000000000c0000004c006f00630061006c00530079007300740065006d0000000000000000000000")
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -750,22 +513,22 @@ func TestRCreateServiceReq(t *testing.T) {
 	}
 
 	req := RCreateServiceRequest{
-        SCContextHandle: handle,
-        ServiceName: NewUnicodeStr(0, "MiscSVC12"),
-        DisplayName: NewUnicodeStr(1, "MiscSVC12"),
-        DesiredAccess: 0x000f01ff,
-        ServiceType: 0x10,
-        StartType: 3,
-        ErrorControl: 0,
-        BinaryPathName: NewUnicodeStr(0, `C:\windows\temp\onedrive.exe`),
-        LoadOrderGroup: nil,
-        TagId: 0,
-        Dependencies: nil,
-        DependSize: 0,
-        ServiceStartName: NewUnicodeStr(2, `LocalSystem`),
-        Password: nil,
-        PwSize: 0,
-    }
+		SCContextHandle:  handle,
+		ServiceName:      NewUnicodeStr(0, "MiscSVC12"),
+		DisplayName:      NewUnicodeStr(1, "MiscSVC12"),
+		DesiredAccess:    0x000f01ff,
+		ServiceType:      0x10,
+		StartType:        3,
+		ErrorControl:     0,
+		BinaryPathName:   NewUnicodeStr(0, `C:\windows\temp\onedrive.exe`),
+		LoadOrderGroup:   nil,
+		TagId:            0,
+		Dependencies:     nil,
+		DependSize:       0,
+		ServiceStartName: NewUnicodeStr(2, `LocalSystem`),
+		Password:         nil,
+		PwSize:           0,
+	}
 
 	buf, err := encoder.Marshal(req)
 	if err != nil {
@@ -773,18 +536,18 @@ func TestRCreateServiceReq(t *testing.T) {
 	}
 
 	if !bytes.Equal(pkt, buf) {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 }
 
 func TestCreateServiceRes(t *testing.T) {
-    pkt, err := hex.DecodeString("0000000000000000e3c4f2041118194eac6a622c8bb6f66c00000000")
+	pkt, err := hex.DecodeString("0000000000000000e3c4f2041118194eac6a622c8bb6f66c00000000")
 	// Simple test to verify that the packet structure is valid
 	if err != nil {
 		t.Fatal(err)
 	}
 
-    handle, err := hex.DecodeString("00000000e3c4f2041118194eac6a622c8bb6f66c")
+	handle, err := hex.DecodeString("00000000e3c4f2041118194eac6a622c8bb6f66c")
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -796,21 +559,21 @@ func TestCreateServiceRes(t *testing.T) {
 	}
 
 	if res.TagId != 0 {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 
 	if res.ReturnCode != 0 {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 
-    if !bytes.Equal(res.ContextHandle, handle) {
-        t.Error("Fail")
-    }
+	if !bytes.Equal(res.ContextHandle, handle) {
+		t.Fatal("Fail")
+	}
 }
 
 func TestDeleteServiceReq(t *testing.T) {
 	// Simple test to verify that the packet structure is valid
-    pkt, err := hex.DecodeString("0000000038a391b8157fcd4198346c9d4d5d4706")
+	pkt, err := hex.DecodeString("0000000038a391b8157fcd4198346c9d4d5d4706")
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -821,8 +584,8 @@ func TestDeleteServiceReq(t *testing.T) {
 	}
 
 	req := RDeleteServiceRequest{
-        ServiceHandle: handle,
-    }
+		ServiceHandle: handle,
+	}
 
 	buf, err := encoder.Marshal(req)
 	if err != nil {
@@ -830,6 +593,6 @@ func TestDeleteServiceReq(t *testing.T) {
 	}
 
 	if !bytes.Equal(pkt, buf) {
-		t.Error("Fail")
+		t.Fatal("Fail")
 	}
 }
